Trait pyo3::buffer::Element

pub unsafe trait Element: Copy {
    fn is_compatible_format(format: &CStr) -> bool;
}

Trait implemented for possible element types of PyBuffer.

Safety

This trait must only be implemented for types which represent valid elements of Python buffers.

Required Methods

fn is_compatible_format(format: &CStr) -> bool

Gets whether the element specified in the format string is potentially compatible. Alignment and size are checked separately from this function.