Function pyo3::callback::handle_panic

pub unsafe fn handle_panic<F, R>(body: F) -> R where
    F: for<'py> FnOnce(Python<'py>) -> PyResult<R> + UnwindSafe,
    R: PyCallbackOutput, 

Variant of the above which does not perform the callback conversion. This allows the callback conversion to be done manually in the case where lifetimes might otherwise cause issue.

For example this pyfunction:

fn foo(&self) -> &Bar {
    &self.bar
}

It is wrapped in proc macros with handle_panic like so:

pyo3::callback::handle_panic(|_py| {
    let _slf = #slf;
    pyo3::callback::convert(_py, #foo)
})

If callback_body was used instead:

pyo3::callback_body!(py, {
    let _slf = #slf;
    #foo
})

Then this will fail to compile, because the result of #foo borrows _slf, but _slf drops when the block passed to the macro ends.