Trait pyo3::callback::PyCallbackOutput

pub trait PyCallbackOutput: Copy {
    const ERR_VALUE: Self;
}

A type which can be the return type of a python C-API callback

Required Associated Constants

const ERR_VALUE: Self

The error value to return to python if the callback raised an exception