⚠️ Internal Docs ⚠️ Not Public API 👉 Official Docs Here

Trait pyo3::class::impl_::PyClassImpl

pub trait PyClassImpl: Sized {
    type Layout: PyLayout<Self>;
    type BaseType: PyTypeInfo + PyTypeObject + PyClassBaseType;
    type ThreadChecker: PyClassThreadChecker<Self>;
    type Inventory: PyClassInventory;

    const DOC: &'static str;
    const IS_GC: bool;
    const IS_BASETYPE: bool;
    const IS_SUBCLASS: bool;
    fn for_each_method_def(_visitor: &mut dyn FnMut(&[PyMethodDefType])) { ... }

    fn for_each_proto_slot(_visitor: &mut dyn FnMut(&[PyType_Slot])) { ... }

    fn get_new() -> Option<newfunc> { ... }

    fn get_alloc() -> Option<allocfunc> { ... }

    fn get_free() -> Option<freefunc> { ... }

    fn dict_offset() -> Option<Py_ssize_t> { ... }

    fn weaklist_offset() -> Option<Py_ssize_t> { ... }
}

Implements the underlying functionality of #[pyclass], assembled by various proc macros.

Users are discouraged from implementing this trait manually; it is a PyO3 implementation detail and may be changed at any time.

Associated Types

type Layout: PyLayout<Self>

Layout

type BaseType: PyTypeInfo + PyTypeObject + PyClassBaseType

Base class

type ThreadChecker: PyClassThreadChecker<Self>

This handles following two situations:

In case T is Send, stub ThreadChecker is used and does nothing. This implementation is used by default. Compile fails if T: !Send.

In case T is !Send, ThreadChecker panics when T is accessed by another thread. This implementation is used when #[pyclass(unsendable)] is given. Panicking makes it safe to expose T: !Send to the Python interpreter, where all objects can be accessed by multiple threads by threading module.

type Inventory: PyClassInventory

This is supported on crate feature multiple-pymethods only.

Associated Constants

const DOC: &'static str

Class doc string

const IS_GC: bool

#[pyclass(gc)]

const IS_BASETYPE: bool

#[pyclass(subclass)]

const IS_SUBCLASS: bool

#[pyclass(extends=…)]

Provided methods

fn for_each_method_def(_visitor: &mut dyn FnMut(&[PyMethodDefType]))

fn for_each_proto_slot(_visitor: &mut dyn FnMut(&[PyType_Slot]))

fn get_new() -> Option<newfunc>

fn get_alloc() -> Option<allocfunc>

fn get_free() -> Option<freefunc>

fn dict_offset() -> Option<Py_ssize_t>

fn weaklist_offset() -> Option<Py_ssize_t>