pyo3::exceptions
Module asyncio

Exceptions defined in Python’s asyncio module.

Structs

CancelledError

A Rust type representing an exception defined in Python code.

IncompleteReadError

A Rust type representing an exception defined in Python code.

InvalidStateError

A Rust type representing an exception defined in Python code.

LimitOverrunError

A Rust type representing an exception defined in Python code.

QueueEmpty

A Rust type representing an exception defined in Python code.

QueueFull

A Rust type representing an exception defined in Python code.

TimeoutError

A Rust type representing an exception defined in Python code.