⚠️ Internal Docs ⚠️ Not Public API 👉 Official Docs Here

Module pyo3::ffi::datetime

This is supported on non-Py_LIMITED_API only.

FFI bindings to the functions and structs defined in datetime.h

This is the unsafe thin wrapper around the CPython C API, and covers the various date and time related objects in the Python datetime standard library module.

A note regarding PyPy (cpyext) support:

Support for PyDateTime_CAPI is limited as of PyPy 7.0.0. DateTime_FromTimestamp and Date_FromTimestamp are currently not supported.

Macros

_PyDateTime_GET_FOLDNon-PyPy

_PyDateTime_GET_HOURNon-PyPy

_PyDateTime_GET_MICROSECONDNon-PyPy

_PyDateTime_GET_MINUTENon-PyPy

_PyDateTime_GET_SECONDNon-PyPy

_PyDateTime_GET_TZINFONon-PyPy

_access_delta_fieldNon-PyPy

_access_fieldNon-PyPy

Structs

PyDateTime_CAPI

PyDateTime_Date

Structure representing a datetime.date

PyDateTime_DateTime

Structure representing a datetime.datetime

PyDateTime_Delta

Structure representing a datetime.timedelta.

PyDateTime_Time

Structure representing a datetime.time.

_PyDateTimeAPI_impl

_PyDateTime_TimeZone_UTC_implNot (PyPy and non-Py_3_8)

Constants

_PyDateTime_DATETIME_DATASIZE

_PyDateTime_DATE_DATASIZE

_PyDateTime_TIME_DATASIZE

Statics

PyDateTimeAPI

Safe wrapper around the Python datetime C-API global. Note that this object differs slightly from the equivalent C object: in C, this is implemented as a static PyDateTime_CAPI *. Here this is implemented as a wrapper which implements Deref to access a reference to a PyDateTime_CAPI object.

PyDateTime_TimeZone_UTCNot (PyPy and non-Py_3_8)

Safe wrapper around the Python C-API global PyDateTime_TimeZone_UTC. This follows a similar strategy as PyDateTimeAPI: the Python datetime C-API will automatically be imported if this type is deferenced.

Functions

PyDateTime_Check⚠

Check if op is a PyDateTimeAPI.DateTimeType or subtype.

PyDateTime_CheckExact⚠

Check if op’s type is exactly PyDateTimeAPI.DateTimeType.

PyDateTime_DATE_GET_FOLD⚠Non-PyPy

Retrieve the fold component of a PyDateTime_DateTime. Returns a signed integer in the interval [0, 1]

PyDateTime_DATE_GET_HOUR⚠Non-PyPy

Retrieve the hour component of a PyDateTime_DateTime. Returns a signed integer in the interval [0, 23]

PyDateTime_DATE_GET_MICROSECOND⚠Non-PyPy

Retrieve the microsecond component of a PyDateTime_DateTime. Returns a signed integer in the interval [0, 999999]

PyDateTime_DATE_GET_MINUTE⚠Non-PyPy

Retrieve the minute component of a PyDateTime_DateTime. Returns a signed integer in the interval [0, 59]

PyDateTime_DATE_GET_SECOND⚠Non-PyPy

Retrieve the second component of a PyDateTime_DateTime. Returns a signed integer in the interval [0, 59]

PyDateTime_DATE_GET_TZINFO⚠Non-PyPy

Retrieve the tzinfo component of a PyDateTime_DateTime. Returns a pointer to a PyObject that should be either NULL or an instance of a datetime.tzinfo subclass.

PyDateTime_DELTA_GET_DAYS⚠Non-PyPy

Retrieve the days component of a PyDateTime_Delta.

PyDateTime_DELTA_GET_MICROSECONDS⚠Non-PyPy

Retrieve the seconds component of a PyDateTime_Delta.

PyDateTime_DELTA_GET_SECONDS⚠Non-PyPy

Retrieve the seconds component of a PyDateTime_Delta.

PyDateTime_FromTimestamp⚠Non-PyPy

PyDateTime_GET_DAY⚠Non-PyPy

Retrieve the day component of a PyDateTime_Date or PyDateTime_DateTime. Returns a signed integer in the interval [1, 31].

PyDateTime_GET_MONTH⚠Non-PyPy

Retrieve the month component of a PyDateTime_Date or PyDateTime_DateTime. Returns a signed integer in the range [1, 12].

PyDateTime_GET_YEAR⚠Non-PyPy

Retrieve the year component of a PyDateTime_Date or PyDateTime_DateTime. Returns a signed integer greater than 0.

PyDateTime_IMPORT⚠

Populates the PyDateTimeAPI object

PyDateTime_TIME_GET_FOLD⚠Non-PyPy

Retrieve the fold component of a PyDateTime_Time. Returns a signed integer in the interval [0, 1]

PyDateTime_TIME_GET_HOUR⚠Non-PyPy

Retrieve the hour component of a PyDateTime_Time. Returns a signed integer in the interval [0, 23]

PyDateTime_TIME_GET_MICROSECOND⚠Non-PyPy

Retrieve the microsecond component of a PyDateTime_DateTime. Returns a signed integer in the interval [0, 999999]

PyDateTime_TIME_GET_MINUTE⚠Non-PyPy

Retrieve the minute component of a PyDateTime_Time. Returns a signed integer in the interval [0, 59]

PyDateTime_TIME_GET_SECOND⚠Non-PyPy

Retrieve the second component of a PyDateTime_DateTime. Returns a signed integer in the interval [0, 59]

PyDateTime_TIME_GET_TZINFO⚠Non-PyPy

Retrieve the tzinfo component of a PyDateTime_Time. Returns a pointer to a PyObject that should be either NULL or an instance of a datetime.tzinfo subclass.

PyDate_Check⚠

Type Check macros

PyDate_CheckExact⚠

Check if op’s type is exactly PyDateTimeAPI.DateType.

PyDate_FromTimestamp⚠Non-PyPy

PyDelta_Check⚠

Check if op is a PyDateTimeAPI.DetaType or subtype.

PyDelta_CheckExact⚠

Check if op’s type is exactly PyDateTimeAPI.DeltaType.

PyTZInfo_Check⚠

Check if op is a PyDateTimeAPI.TZInfoType or subtype.

PyTZInfo_CheckExact⚠

Check if op’s type is exactly PyDateTimeAPI.TZInfoType.

PyTime_Check⚠

Check if op is a PyDateTimeAPI.TimeType or subtype.

PyTime_CheckExact⚠

Check if op’s type is exactly PyDateTimeAPI.TimeType.