Module pyo3::freelist

Free allocation list

Structs

FreeList

Enums

Slot

Traits

PyObjectWithFreeList
Implementing this trait for custom class adds free allocation list to class. The performance improvement applies to types that are often created and deleted in a row, so that they can benefit from a freelist.