⚠️ Internal Docs ⚠️ Not Public API 👉 Official Docs Here

Module pyo3::hashbrown

This is supported on crate feature hashbrown only.

Conversions to and from hashbrown’s HashMap and HashSet.

Setup

To use this feature, add this to your Cargo.toml:

[dependencies]
hashbrown = "*"
pyo3 = { version = "0.15.1", features = ["hashbrown"] }

Note that you must use compatible versions of hashbrown and PyO3. The required hashbrown version may vary based on the version of PyO3.