⚠️ Internal Docs ⚠️ Not Public API 👉 Official Docs Here

Struct pyo3::impl_::extract_argument::FunctionDescription

pub struct FunctionDescription {
    pub cls_name: Option<&'static str>,
    pub func_name: &'static str,
    pub positional_parameter_names: &'static [&'static str],
    pub positional_only_parameters: usize,
    pub required_positional_parameters: usize,
    pub keyword_only_parameters: &'static [KeywordOnlyParameterDescription],
}

Function argument specification for a #[pyfunction] or #[pymethod].

Fields
cls_name: Option<&'static str>func_name: &'static strpositional_parameter_names: &'static [&'static str]positional_only_parameters: usizerequired_positional_parameters: usizekeyword_only_parameters: &'static [KeywordOnlyParameterDescription]
Implementations

impl FunctionDescription

fn full_name(&self) -> String

pub unsafe fn extract_arguments_fastcall<'py, V, K>(
&self,
py: Python<'py>,
args: *const *mut PyObject,
nargs: Py_ssize_t,
kwnames: *mut PyObject,
output: &mut [Option<&'py PyAny>]
) -> PyResult<(V::Varargs, K::Varkeywords)> where
V: VarargsHandler<'py>,
K: VarkeywordsHandler<'py>,

This is supported on non-Py_LIMITED_API only.

Equivalent of extract_arguments_tuple_dict which uses the Python C-API “fastcall” convention.

Safety

args must be a pointer to a C-style array of valid ffi::PyObject pointers.

kwnames must be a pointer to a PyTuple, or NULL.

nargs + kwnames.len() is the total length of the args array.

pub unsafe fn extract_arguments_tuple_dict<'py, V, K>(
&self,
py: Python<'py>,
args: *mut PyObject,
kwargs: *mut PyObject,
output: &mut [Option<&'py PyAny>]
) -> PyResult<(V::Varargs, K::Varkeywords)> where
V: VarargsHandler<'py>,
K: VarkeywordsHandler<'py>,

Extracts the args and kwargs provided into output, according to this function definition.

output must have the same length as this function has positional and keyword-only parameters (as per the positional_parameter_names and keyword_only_parameters respectively).

Unexpected, duplicate or invalid arguments will cause this function to return TypeError.

Safety

args must be a pointer to a PyTuple.

kwargs must be a pointer to a PyDict, or NULL.

fn find_keyword_parameter_in_positional(
&self,
kwarg_name: &str
) -> Option<usize>

fn find_keyword_parameter_in_keyword_only(
&self,
kwarg_name: &str
) -> Option<usize>

fn ensure_no_missing_required_positional_arguments(
&self,
output: &[Option<&PyAny>],
positional_args_provided: usize
) -> PyResult<()>

fn ensure_no_missing_required_keyword_arguments(
&self,
output: &[Option<&PyAny>]
) -> PyResult<()>

fn too_many_positional_arguments(&self, args_provided: usize) -> PyErr

fn multiple_values_for_argument(&self, argument: &str) -> PyErr

fn unexpected_keyword_argument(&self, argument: &PyAny) -> PyErr

fn positional_only_keyword_arguments(&self, parameter_names: &[&str]) -> PyErr

fn missing_required_arguments(
&self,
argument_type: &str,
parameter_names: &[&str]
) -> PyErr

fn missing_required_keyword_arguments(
&self,
keyword_outputs: &[Option<&PyAny>]
) -> PyErr

fn missing_required_positional_arguments(
&self,
output: &[Option<&PyAny>]
) -> PyErr