Trait pyo3::impl_::extract_argument::PyFunctionArgument

pub trait PyFunctionArgument<'a, 'py>: Sized + 'a {
    type Holder: FunctionArgumentHolder;

    fn extract(obj: &'py PyAny, holder: &'a mut Self::Holder) -> PyResult<Self>;
}

A trait which is used to help PyO3 macros extract function arguments.

#[pyclass] structs need to extract as PyRef<T> and PyRefMut<T> wrappers rather than extracting &T and &mut T directly. The Holder type is used to hold these temporary wrappers - the way the macro is constructed, these wrappers will be dropped as soon as the pyfunction call ends.

There exists a trivial blanket implementation for T: FromPyObject with Holder = ().

Required Associated Types

type Holder: FunctionArgumentHolder

Required Methods

fn extract(obj: &'py PyAny, holder: &'a mut Self::Holder) -> PyResult<Self>