⚠️ Internal Docs ⚠️ Not Public API 👉 Official Docs Here

Trait pyo3::impl_::extract_argument::VarargsHandler

pub trait VarargsHandler<'py> {
    type Varargs;
    fn handle_varargs_fastcall(
        py: Python<'py>, 
        varargs: &[Option<&PyAny>], 
        function_description: &FunctionDescription
    ) -> PyResult<Self::Varargs>;

    fn handle_varargs_tuple(
        args: &'py PyTuple, 
        function_description: &FunctionDescription
    ) -> PyResult<Self::Varargs>;
}

A trait used to control whether to accept varargs in FunctionDescription::extract_argument_(method) functions.

Associated Types

type Varargs

Required methods

fn handle_varargs_fastcall(
py: Python<'py>,
varargs: &[Option<&PyAny>],
function_description: &FunctionDescription
) -> PyResult<Self::Varargs>

Called by FunctionDescription::extract_arguments_fastcall with any additional arguments.

fn handle_varargs_tuple(
args: &'py PyTuple,
function_description: &FunctionDescription
) -> PyResult<Self::Varargs>

Called by FunctionDescription::extract_arguments_tuple_dict with the original tuple.

Additional arguments are those in the tuple slice starting from function_description.positional_parameter_names.len().