âš ï¸ Internal Docs âš ï¸ Not Public API ğŸ‘‰ Official Docs Here

Struct pyo3::impl_::pyclass::ThreadCheckerInherited

pub struct ThreadCheckerInherited<T: Send, U: PyClassBaseType>(PhantomData<T>, U::ThreadChecker);

Thread checker for types that have Send and extends=.... Ensures that T: Send and the parent is not accessed by another thread.

Tuple Fields
0: PhantomData<T>1: U::ThreadChecker