Module pyo3::prelude

PyO3’s prelude.

The purpose of this module is to alleviate imports of many commonly used items of the PyO3 crate by adding a glob import to the top of pyo3 heavy modules:

use pyo3::prelude::*;

Re-exports

pub use crate::conversion::FromPyObject;

pub use crate::conversion::IntoPy;

pub use crate::conversion::IntoPyPointer;

pub use crate::conversion::PyTryFrom;

pub use crate::conversion::PyTryInto;

pub use crate::conversion::ToPyObject;

pub use crate::marker::Python;

pub use crate::pycell::PyCell;

pub use crate::pycell::PyRef;

pub use crate::pycell::PyRefMut;

pub use crate::pyclass_init::PyClassInitializer;

Macros

wrap_pyfunction

Wraps a Rust function annotated with #[pyfunction].

Structs

GILGuard

RAII type that represents the Global Interpreter Lock acquisition.

Py

A GIL-independent reference to an object allocated on the Python heap.

PyAny

Represents any Python object.

PyErr

Represents a Python exception.

PyModule

Represents a Python module object.

Type Definitions

PyObject

A commonly-used alias for Py<PyAny>.

PyResult

Represents the result of a Python call.

Attribute Macros

pyclass

A proc macro used to expose Rust structs and fieldless enums as Python objects.

pyfunction

A proc macro used to expose Rust functions to Python.

pymethods

A proc macro used to expose methods to Python.

pymodule

A proc macro used to implement Python modules.

pyproto

A proc macro used to implement Python’s dunder methods.

Derive Macros

FromPyObject