pyo3::pycell
Struct PyRefMut

pub struct PyRefMut<'p, T: PyClass<Frozen = False>> { /* private fields */ }

A wrapper type for a mutably borrowed value from a [Bound<'py, T>].

See the module-level documentation for more information.

Implementations

impl<'p, T: PyClass<Frozen = False>> PyRefMut<'p, T>

pub fn py(&self) -> Python<'p>

Returns a Python token that is bound to the lifetime of the PyRefMut.

impl<'py, T: PyClass<Frozen = False>> PyRefMut<'py, T>

pub fn as_ptr(&self) -> *mut PyObject

Returns the raw FFI pointer represented by self.

Safety

Callers are responsible for ensuring that the pointer does not outlive self.

The reference is borrowed; callers should not decrease the reference count when they are finished with the pointer.

pub fn into_ptr(self) -> *mut PyObject

Returns an owned raw FFI pointer represented by self.

Safety

The reference is owned; when finished the caller should either transfer ownership of the pointer or decrease the reference count (e.g. with pyo3::ffi::Py_DecRef).

impl<'p, T, U> PyRefMut<'p, T>
where T: PyClass<BaseType = U, Frozen = False>, U: PyClass<Frozen = False>,

pub fn into_super(self) -> PyRefMut<'p, U>

Gets a PyRef<T::BaseType>.

See PyRef::into_super for more.

pub fn as_super(&mut self) -> &mut PyRefMut<'p, U>

Borrows a mutable reference to PyRefMut<T::BaseType>.

With the help of this method, you can mutate attributes and call mutating methods on the superclass without consuming the PyRefMut<T>. This method can also be chained to access the super-superclass (and so on).

See PyRef::as_super for more.