Trait pyo3::pycell::PyCellLayout

pub trait PyCellLayout<T>: PyLayout<T> {
    fn ensure_threadsafe(&self);
    unsafe fn tp_dealloc(slf: *mut PyObject, py: Python<'_>);
}

Required Methods

fn ensure_threadsafe(&self)

unsafe fn tp_dealloc(slf: *mut PyObject, py: Python<'_>)

Implementation of tp_dealloc.

Safety

slf must be a valid pointer to an instance of a T or a subclass.

slf must not be used after this call (as it will be freed).