⚠️ Internal Docs ⚠️ Not Public API 👉 Official Docs Here

Module pyo3::pyclass

PyClass and related traits.

Structs

PyClassInfo

Enums

CompareOp

Operators for the __richcmp__ method

IterANextOutput

Output of __anext__.

IterNextOutput

Output of __next__ which can either yield the next value in the iteration, or return a value to raise StopIteration in Python.

Constants

PY_GET_SET_DEF_INIT

Traits

PyClass

If PyClass is implemented for a Rust type T, then we can use T in the Python world, via PyCell.

Functions

create_type_object

create_type_object_impl⚠

into_raw

method_defs_to_pyclass_info

no_constructor_defined⚠

Default new implementation

py_class_doc

py_class_flags

py_class_qualified_name

tp_init_additional⚠Non-Py_LIMITED_API and non-Py_3_10

Additional type initializations necessary before Python 3.10

type_object_creation_failed

Type Definitions

PyIterANextOutput

An IterANextOutput of Python objects.

PyIterNextOutput