Struct pyo3::pyclass::PyClassShell

#[repr(C)]
pub struct PyClassShell<T: PyClass> { /* fields omitted */ }

PyClassShell represents the concrete layout of T: PyClass when it is converted to a Python class.

You can use it to test your #[pyclass] correctly works.

#[pyclass]
struct Book {
    #[pyo3(get)]
    name: &'static str,
    author: &'static str,
}
let gil = Python::acquire_gil();
let py = gil.python();
let book = Book {
    name: "The Man in the High Castle",
    author: "Philip Kindred Dick",
};
let book_shell = PyClassShell::new_ref(py, book).unwrap();
py_run!(py, book_shell, "assert book_shell.name[-6:] == 'Castle'");

Methods

impl<T: PyClass> PyClassShell<T>

pub fn new_ref(
py: Python,
value: impl Into<PyClassInitializer<T>>
) -> PyResult<&Self> where
<T::BaseType as PyTypeInfo>::ConcreteLayout: PyObjectSizedLayout<T::BaseType>,

Make new PyClassShell on the Python heap and returns the reference of it.

pub fn new_mut(
py: Python,
value: impl Into<PyClassInitializer<T>>
) -> PyResult<&mut Self> where
<T::BaseType as PyTypeInfo>::ConcreteLayout: PyObjectSizedLayout<T::BaseType>,

Make new PyClassShell on the Python heap and returns the mutable reference of it.

pub fn get_super(&self) -> &<T::BaseType as PyTypeInfo>::ConcreteLayout

Get the reference of base object.

pub fn get_super_mut(
&mut self
) -> &mut <T::BaseType as PyTypeInfo>::ConcreteLayout

Get the mutable reference of base object.