Trait pyo3::pyclass::PyClassAlloc

pub trait PyClassAlloc: PyTypeInfo + Sized {
    unsafe fn new(
        py: Python<'_>, 
        subtype: *mut PyTypeObject
    ) -> *mut Self::Layout { ... }

    unsafe fn dealloc(py: Python<'_>, self_: *mut Self::Layout) { ... }
}

This trait enables custom tp_new/tp_dealloc implementations for T: PyClass.

Provided methods

unsafe fn new(py: Python<'_>, subtype: *mut PyTypeObject) -> *mut Self::Layout

Allocate the actual field for #[pyclass].

Safety

This function must return a valid pointer to the Python heap.

unsafe fn dealloc(py: Python<'_>, self_: *mut Self::Layout)

Deallocate #[pyclass] on the Python heap.

Safety

self_ must be a valid pointer to the Python heap.