Struct pyo3::ManagedPyRef

#[repr(transparent)]pub struct ManagedPyRef<'p, T: ToPyObject + ?Sized> { /* fields omitted */ }

Reference to a converted ToPyObject.

Many methods want to take anything that can be converted into a Python object. This type takes care of both types types that are already Python object (i.e. implement AsPyPointer) and those that don't (i.e. ToPyObject types). For the AsPyPointer types, we just use the borrowed pointer, which is a lot faster and simpler than creating a new extra object. The remaning ToPyObject types are converted to Python objects, the owned pointer is stored and decref'd on drop.

Example

use pyo3::ffi;
use pyo3::{ToPyObject, AsPyPointer, PyNativeType, ManagedPyRef};
use pyo3::types::{PyDict, PyAny};

pub fn get_dict_item<'p>(dict: &'p PyDict, key: &impl ToPyObject) -> Option<&'p PyAny> {
    let key = ManagedPyRef::from_to_pyobject(dict.py(), key);
    unsafe {
        dict.py().from_borrowed_ptr_or_opt(ffi::PyDict_GetItem(dict.as_ptr(), key.as_ptr()))
    }
}

Implementations

impl<'p, T: ToPyObject> ManagedPyRef<'p, T>

This should eventually be replaced with a generic IntoPy trait impl by figuring out the correct lifetime annotation to make the compiler happy.

pub fn from_to_pyobject(py: Python<'p>, to_pyobject: &T) -> Self