Struct pyo3::PyObject

pub struct PyObject(_);

Safe wrapper around unsafe *mut ffi::PyObject pointer.

Methods

impl PyObject

unsafe fn from_owned_ptr(_py: Python, ptr: *mut PyObject) -> PyObject

Creates a PyObject instance for the given FFI pointer. This moves ownership over the pointer into the PyObject. Undefined behavior if the pointer is NULL or invalid.

unsafe fn from_owned_ptr_or_panic(py: Python, ptr: *mut PyObject) -> PyObject

Creates a PyObject instance for the given FFI pointer. Panics if the pointer is null. Undefined behavior if the pointer is invalid.

fn from_owned_ptr_or_err(py: Python, ptr: *mut PyObject) -> PyResult<PyObject>

Construct PyObject from the result of a Python FFI call that returns a new reference (owned pointer). Returns Err(PyErr) if the pointer is null.

fn from_owned_ptr_or_opt(py: Python, ptr: *mut PyObject) -> Option<PyObject>

Construct PyObject from the result of a Python FFI call that returns a new reference (owned pointer). Returns None if the pointer is null.

unsafe fn from_borrowed_ptr(_py: Python, ptr: *mut PyObject) -> PyObject

Creates a PyObject instance for the given Python FFI pointer. Calls Py_INCREF() on the ptr. Undefined behavior if the pointer is NULL or invalid.

fn from_borrowed_ptr_or_err(
py: Python,
ptr: *mut PyObject
) -> PyResult<PyObject>

Creates a PyObject instance for the given Python FFI pointer. Calls Py_INCREF() on the ptr. Returns Err(PyErr) if the pointer is null.

fn from_borrowed_ptr_or_opt(py: Python, ptr: *mut PyObject) -> Option<PyObject>

Creates a PyObject instance for the given Python FFI pointer. Calls Py_INCREF() on the ptr. Returns None if the pointer is null.

unsafe fn borrow_from_owned_ptr_slice(ptr: &[*mut PyObject]) -> &[PyObject]

Transmutes a slice of owned FFI pointers to &[PyObject]. Undefined behavior if any pointer in the slice is NULL or invalid.

fn get_refcnt(&self) -> isize

Gets the reference count of the ffi::PyObject pointer.

fn clone_ref(&self, py: Python) -> Self

Clone self, Calls Py_INCREF() on the ptr.

fn is_none(&self) -> bool

Returns whether the object is considered to be None. This is equivalent to the Python expression: 'is None'

fn is_true(&self, py: Python) -> PyResult<bool>

Returns whether the object is considered to be true. This is equivalent to the Python expression: 'not not self'

fn cast_as<D>(&self, py: Python) -> Option<&D> where
D: PyDowncastFrom,

Casts the PyObject to a concrete Python object type.

fn extract<'p, D>(&'p self, py: Python) -> PyResult<D> where
D: FromPyObject<'p>,

Extracts some type from the Python object. This is a wrapper function around FromPyObject::extract().

fn getattr<N>(&self, py: Python, attr_name: N) -> PyResult<PyObject> where
N: ToPyObject,

Retrieves an attribute value. This is equivalent to the Python expression 'self.attr_name'.

fn call<A>(
&self,
py: Python,
args: A,
kwargs: Option<&PyDict>
) -> PyResult<PyObject> where
A: IntoPyTuple,

Calls the object. This is equivalent to the Python expression: 'self(*args, **kwargs)'

fn call_method<A>(
&self,
py: Python,
name: &str,
args: A,
kwargs: Option<&PyDict>
) -> PyResult<PyObject> where
A: IntoPyTuple,

Calls a method on the object. This is equivalent to the Python expression: 'self.name(*args, **kwargs)'