Struct pyo3::PyTuple

pub struct PyTuple(_);

Represents a Python tuple object.

Methods

impl PyTuple

pub fn new<T: ToPyObject>(py: Python, elements: &[T]) -> Py<PyTuple>

Construct a new tuple with the given elements.

pub fn empty(_py: Python) -> Py<PyTuple>

Retrieves the empty tuple.

pub fn len(&self) -> usize

Gets the length of the tuple.

pub fn slice(&self, low: isize, high: isize) -> Py<PyTuple>

Take a slice of the tuple pointed to by p from low to high and return it as a new tuple.

pub fn split_from(&self, low: isize) -> Py<PyTuple>

Take a slice of the tuple pointed to by p from low and return it as a new tuple.

pub fn get_item(&self, index: usize) -> &PyObjectRef

Gets the item at the specified index.

Panics if the index is out of range.

pub fn as_slice(&self) -> &[PyObject]

pub fn iter(&self) -> PyTupleIterator

Returns an iterator over the tuple items.