Struct pyo3::PyUnicode

pub struct PyUnicode(_);

Represents a Python string.

Methods

impl PyString

pub fn new(_py: Python, s: &str) -> Py<PyString>

Creates a new Python string object.

Panics if out of memory.

pub fn from_object<'p>(
src: &'p PyObjectRef,
encoding: &str,
errors: &str
) -> PyResult<&'p PyString>

pub fn data(&self) -> PyStringData

Gets the python string data in its underlying representation.

pub fn to_string(&self) -> PyResult<Cow<str>>

Convert the PyString into a Rust string.

Returns a UnicodeDecodeError if the input is not valid unicode (containing unpaired surrogates).

pub fn to_string_lossy(&self) -> Cow<str>

Convert the PyString into a Rust string.

Unpaired surrogates invalid UTF-8 sequences are replaced with U+FFFD REPLACEMENT CHARACTER.