Module pyo3::type_object

Python type object information

Structs

PyRawObject
Special object that is used for python object creation. pyo3 library automatically creates this object for class __new__ method. Behavior is undefined if constructor of custom class does not initialze instance of PyRawObject with rust value with init method. Calling of __new__ method of base class is developer's responsibility.

Constants

PY_TYPE_FLAG_BASETYPE
Type object can be used as the base type of another type

PY_TYPE_FLAG_DICT
The instances of this type have a dictionary containing instance variables

PY_TYPE_FLAG_GC
type object supports python GC

PY_TYPE_FLAG_WEAKREF
Type object supports python weak references

Traits

PyObjectAlloc
A Python object allocator that is usable as a base type for #[pyclass]

PyTypeCreate
Python object types that can be instanciated with Self::create()

PyTypeInfo
Python type information.

PyTypeObject
Python object types that have a corresponding type object.

Functions

initialize_type
Register new type in python object system.