Trait pyo3::type_object::PyLayout

pub unsafe trait PyLayout<T: PyTypeInfo> {
    const IS_NATIVE_TYPE: bool;
    fn get_super(&mut self) -> Option<&mut T::BaseLayout> { ... }

    unsafe fn py_init(&mut self, _value: T) { ... }

    unsafe fn py_drop(&mut self, _py: Python<'_>) { ... }
}

T: PyLayout<U> represents that T is a concrete representaion of U in Python heap. E.g., PyCell is a concrete representaion of all pyclasses, and ffi::PyObject is of PyAny.

This trait is intended to be used internally.

Associated Constants

const IS_NATIVE_TYPE: bool

Provided methods

fn get_super(&mut self) -> Option<&mut T::BaseLayout>

unsafe fn py_init(&mut self, _value: T)

unsafe fn py_drop(&mut self, _py: Python<'_>)