pyo3::type_object

Trait PySizedLayout

pub trait PySizedLayout<T>: PyLayout<T> + Sized { }

T: PySizedLayout<U> represents that T is not a instance of PyVarObject. In addition, that T is a concrete representation of U.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.