Trait pyo3::typeob::PyObjectAlloc

pub trait PyObjectAlloc: PyTypeInfo + Sized {
    unsafe fn alloc(_py: Python) -> PyResult<*mut PyObject> { ... }

    unsafe fn dealloc(py: Python, obj: *mut PyObject) { ... }

    unsafe fn drop(py: Python, obj: *mut PyObject) { ... }
}

A Python object allocator that is usable as a base type for #[pyclass]

All native types and all #[pyclass] types use the default functions, while PyObjectWithFreeList gets a special version.

Provided methods

unsafe fn alloc(_py: Python) -> PyResult<*mut PyObject>

unsafe fn dealloc(py: Python, obj: *mut PyObject)

Calls the rust destructor for the object and frees the memory (usually by calling ptr->ob_type->tp_free). This function is used as tp_dealloc implementation.

unsafe fn drop(py: Python, obj: *mut PyObject)

Calls the rust destructor for the object.

Loading content...