Struct pyo3::types::num::PyLong

#[repr(transparent)]
pub struct PyLong(PyAny);

Represents a Python int object.

You can usually avoid directly working with this type by using ToPyObject and extract with the primitive Rust integer types.

Tuple Fields
0: PyAny