pyo3::types
Struct PyCFunction

pub struct PyCFunction(/* private fields */);

Represents a builtin Python function object.

Values of this type are accessed via PyO3โs smart pointers, e.g. as Py<PyCFunction> or Bound<'py, PyCFunction>.

Implementations

impl PyCFunction

pub fn new_with_keywords<'py>( py: Python<'py>, fun: PyCFunctionWithKeywords, name: &'static CStr, doc: &'static CStr, module: Option<&Bound<'py, PyModule>>, ) -> PyResult<Bound<'py, Self>>

Create a new built-in function with keywords (*args and/or **kwargs).

To create name and doc static strings on Rust versions older than 1.77 (which added cโโ literals), use the c_str! macro.

pub fn new_with_keywords_bound<'py>( py: Python<'py>, fun: PyCFunctionWithKeywords, name: &'static CStr, doc: &'static CStr, module: Option<&Bound<'py, PyModule>>, ) -> PyResult<Bound<'py, Self>>

๐Deprecated since 0.23.0: renamed to PyCFunction::new_with_keywords

Deprecated name for PyCFunction::new_with_keywords.

pub fn new<'py>( py: Python<'py>, fun: PyCFunction, name: &'static CStr, doc: &'static CStr, module: Option<&Bound<'py, PyModule>>, ) -> PyResult<Bound<'py, Self>>

Create a new built-in function which takes no arguments.

To create name and doc static strings on Rust versions older than 1.77 (which added cโโ literals), use the c_str! macro.

pub fn new_bound<'py>( py: Python<'py>, fun: PyCFunction, name: &'static CStr, doc: &'static CStr, module: Option<&Bound<'py, PyModule>>, ) -> PyResult<Bound<'py, Self>>

๐Deprecated since 0.23.0: renamed to PyCFunction::new

Deprecated name for PyCFunction::new.

pub fn new_closure<'py, F, R>( py: Python<'py>, name: Option<&'static CStr>, doc: Option<&'static CStr>, closure: F, ) -> PyResult<Bound<'py, Self>>
where F: Fn(&Bound<'_, PyTuple>, Option<&Bound<'_, PyDict>>) -> R + Send + 'static, for<'p> R: IntoPyCallbackOutput<'p, *mut PyObject>,

Create a new function from a closure.

Examples

Python::with_gil(|py| {
    let add_one = |args: &Bound<'_, PyTuple>, _kwargs: Option<&Bound<'_, PyDict>>| -> PyResult<_> {
        let i = args.extract::<(i64,)>()?.0;
        Ok(i+1)
    };
    let add_one = PyCFunction::new_closure(py, None, None, add_one).unwrap();
    py_run!(py, add_one, "assert add_one(42) == 43");
});

pub fn new_closure_bound<'py, F, R>( py: Python<'py>, name: Option<&'static CStr>, doc: Option<&'static CStr>, closure: F, ) -> PyResult<Bound<'py, Self>>
where F: Fn(&Bound<'_, PyTuple>, Option<&Bound<'_, PyDict>>) -> R + Send + 'static, for<'p> R: IntoPyCallbackOutput<'p, *mut PyObject>,

๐Deprecated since 0.23.0: renamed to PyCFunction::new_closure

Deprecated name for PyCFunction::new_closure.