pyo3::types
Struct PyFrozenSetBuilder

pub struct PyFrozenSetBuilder<'py> { /* private fields */ }

Allows building a Python frozenset one item at a time

Implementations

impl<'py> PyFrozenSetBuilder<'py>

pub fn new(py: Python<'py>) -> PyResult<PyFrozenSetBuilder<'py>>

Create a new FrozenSetBuilder. Since this allocates a PyFrozenSet internally it may panic when running out of memory.

pub fn add<K>(&mut self, key: K) -> PyResult<()>
where K: IntoPyObject<'py>,

Adds an element to the set.

pub fn finalize(self) -> Bound<'py, PyFrozenSet>

Finish building the set and take ownership of its current value

pub fn finalize_bound(self) -> Bound<'py, PyFrozenSet>

๐Deprecated since 0.23.0: renamed to PyFrozenSetBuilder::finalize

Deprecated name for PyFrozenSetBuilder::finalize.