Struct pyo3::types::PyInt

#[repr(transparent)]

pub struct PyInt(_);

Represents a Python int object.

You can usually avoid directly working with this type by using ToPyObject and extract with the primitive Rust integer types.