Struct pyo3::types::PyModule

#[repr(transparent)]

pub struct PyModule(_);

Represents a Python module object.

Methods

impl PyModule

pub fn new<'p>(py: Python<'p>, name: &str) -> PyResult<&'p PyModule>

Create a new module object with the __name__ attribute set to name.

pub fn import<'p>(py: Python<'p>, name: &str) -> PyResult<&'p PyModule>

Import the Python module with the specified name.

pub fn from_code<'p>(
py: Python<'p>,
code: &str,
file_name: &str,
module_name: &str
) -> PyResult<&'p PyModule>

Loads the python code specified into a new module 'code' is the raw Python you want to load into the module 'file_name' is the file name to associate with the module (this is used when Python reports errors, for example) 'module_name' is the name to give the module

pub fn dict(&self) -> &PyDict

Return the dictionary object that implements module's namespace; this object is the same as the __dict__ attribute of the module object.

pub fn name(&self) -> PyResult<&str>

Gets the module name.

May fail if the module does not have a __name__ attribute.

pub fn filename(&self) -> PyResult<&str>

Gets the module filename.

May fail if the module does not have a __file__ attribute.

pub fn call<A>(
&self,
name: &str,
args: A,
kwargs: Option<&PyDict>
) -> PyResult<&PyObjectRef> where
A: IntoPyTuple,

Calls a function in the module. This is equivalent to the Python expression: getattr(module, name)(*args, **kwargs)

pub fn call0(&self, name: &str) -> PyResult<&PyObjectRef>

Calls a function in the module. This is equivalent to the Python expression: getattr(module, name)()

pub fn call1<A>(&self, name: &str, args: A) -> PyResult<&PyObjectRef> where
A: IntoPyTuple,

Calls a function in the module. This is equivalent to the Python expression: getattr(module, name)(*args)

pub fn get(&self, name: &str) -> PyResult<&PyObjectRef>

Gets a member from the module. This is equivalent to the Python expression: getattr(module, name)

pub fn add<V>(&self, name: &str, value: V) -> PyResult<()> where
V: ToPyObject,

Adds a member to the module.

This is a convenience function which can be used from the module's initialization function.

pub fn add_class<T>(&self) -> PyResult<()> where
T: PyTypeInfo,

Adds a new extension type to the module.

This is a convenience function that initializes the class, sets new_type.__module__ to this module's name, and adds the type to this module.

pub fn add_function(&self, wrapper: &dyn Fn(Python) -> PyObject) -> PyResult<()>

Adds a function to a module, using the functions name as name.

Use this together with the#[pyfunction] and wrap_function! macro.

m.add_function(wrap_function!(double));

You can also add a function with a custom name using add:

m.add("also_double", wrap_function!(double)(py));