Struct pyo3::types::PySequence

#[repr(transparent)]pub struct PySequence(_, _);

Represents a reference to a Python object supporting the sequence protocol.

Implementations

impl PySequence

pub fn len(&self) -> PyResult<isize>

Returns the number of objects in sequence.

This is equivalent to the Python expression len(self).

pub fn is_empty(&self) -> PyResult<bool>

pub fn concat(&self, other: &PySequence) -> PyResult<&PySequence>

Returns the concatenation of self and other.

This is equivalent to the Python expression self + other.

pub fn repeat(&self, count: isize) -> PyResult<&PySequence>

Returns the result of repeating a sequence object count times.

This is equivalent to the Python expression self * count. NB: Python accepts negative counts; it returns an empty Sequence.

pub fn in_place_concat(&self, other: &PySequence) -> PyResult<()>

Concatenates self and other in place.

This is equivalent to the Python statement self += other.

pub fn in_place_repeat(&self, count: isize) -> PyResult<()>

Repeats the sequence object count times and updates self.

This is equivalent to the Python statement self *= count. NB: Python accepts negative counts; it empties the Sequence.

pub fn get_item(&self, index: isize) -> PyResult<&PyAny>

Returns the indexth element of the Sequence.

This is equivalent to the Python expression self[index].

pub fn get_slice(&self, begin: isize, end: isize) -> PyResult<&PyAny>

Returns the slice of sequence object between begin and end.

This is equivalent to the Python expression self[begin:end].

pub fn set_item<I>(&self, i: isize, item: I) -> PyResult<()> where
I: ToBorrowedObject,

Assigns object item to the ith element of self.

This is equivalent to the Python statement self[i] = v.

pub fn del_item(&self, i: isize) -> PyResult<()>

Deletes the ith element of self.

This is equivalent to the Python statement del self[i].

pub fn set_slice(&self, i1: isize, i2: isize, v: &PyAny) -> PyResult<()>

Assigns the sequence v to the slice of self from i1 to i2.

This is equivalent to the Python statement self[i1:i2] = v.

pub fn del_slice(&self, i1: isize, i2: isize) -> PyResult<()>

Deletes the slice from i1 to i2 from self.

This is equivalent to the Python statement del self[i1:i2].

pub fn count<V>(&self, value: V) -> PyResult<usize> where
V: ToBorrowedObject,

Returns the number of occurrences of value in self, that is, return the number of keys for which self[key] == value.

pub fn contains<V>(&self, value: V) -> PyResult<bool> where
V: ToBorrowedObject,

Determines if self contains value.

This is equivalent to the Python expression value in self.

pub fn index<V>(&self, value: V) -> PyResult<usize> where
V: ToBorrowedObject,

Returns the first index i for which self[i] == value.

This is equivalent to the Python expression self.index(value).

pub fn list(&self) -> PyResult<&PyList>

Returns a fresh list based on the Sequence.

pub fn tuple(&self) -> PyResult<&PyTuple>

Returns a fresh tuple based on the Sequence.