pyo3::types
Struct PySet

pub struct PySet(/* private fields */);

Represents a Python set.

Values of this type are accessed via PyO3โs smart pointers, e.g. as Py<PySet> or Bound<'py, PySet>.

For APIs available on set objects, see the PySetMethods trait which is implemented for Bound<'py, PySet>.

Implementations

impl PySet

pub fn new<'py, T>( py: Python<'py>, elements: impl IntoIterator<Item = T>, ) -> PyResult<Bound<'py, PySet>>
where T: IntoPyObject<'py>,

Creates a new set with elements from the given slice.

Returns an error if some element is not hashable.

pub fn new_bound<'a, 'p, T: ToPyObject + 'a>( py: Python<'p>, elements: impl IntoIterator<Item = &'a T>, ) -> PyResult<Bound<'p, PySet>>

๐Deprecated since 0.23.0: renamed to PySet::new

Deprecated name for PySet::new.

pub fn empty(py: Python<'_>) -> PyResult<Bound<'_, PySet>>

Creates a new empty set.

pub fn empty_bound(py: Python<'_>) -> PyResult<Bound<'_, PySet>>

๐Deprecated since 0.23.0: renamed to PySet::empty

Deprecated name for PySet::empty.