Struct pyo3::types::PySlice

#[repr(transparent)]

pub struct PySlice(_);

Represents a Python slice.

Only c_long indeces supprted at the moment by PySlice object.

Methods

impl PySlice

pub fn new(py: Python, start: isize, stop: isize, step: isize) -> &PySlice

Construct a new slice with the given elements.

pub fn indices(&self, length: c_long) -> PyResult<PySliceIndices>

Retrieve the start, stop, and step indices from the slice object slice assuming a sequence of length length, and store the length of the slice in slicelength.