pyo3

Module class

Old module which contained some implementation details of the #[pyproto] module.

Prefer using the same content from pyo3::pyclass, e.g. use pyo3::pyclass::CompareOp instead of use pyo3::class::basic::CompareOp.

For compatibility reasons this has not yet been removed, however will be done so once https://github.com/rust-lang/rust/issues/30827 is resolved.

Re-exports

pub use self::gc::PyTraverseError;

pub use self::gc::PyVisit;

Modules

basic

Old module which contained some implementation details of the #[pyproto] module.

gc

Old module which contained some implementation details of the #[pyproto] module.

iter

Old module which contained some implementation details of the #[pyproto] module.

pyasync

Old module which contained some implementation details of the #[pyproto] module.