Module pyo3::class

Python object protocols

Re-exports

pub use self::basic::PyObjectProtocol;

Deprecated

pub use self::buffer::PyBufferProtocol;

Deprecated

pub use self::descr::PyDescrProtocol;

Deprecated

pub use self::gc::PyGCProtocol;

Deprecated

pub use self::iter::PyIterProtocol;

Deprecated

pub use self::mapping::PyMappingProtocol;

Deprecated

pub use self::number::PyNumberProtocol;

Deprecated

pub use self::pyasync::PyAsyncProtocol;

Deprecated

pub use self::sequence::PySequenceProtocol;

Deprecated

Modules

basic

Basic Python Object customization

buffer

Represent Python Buffer protocol implementation

descr

Python Description Interface

gc

Python GC support

iter

Python Iterator Interface. Trait and support implementation for implementing iterators

mapping

Python Mapping Interface Trait and support implementation for implementing mapping support

number

Python Number Interface Trait and support implementation for implementing number protocol

pyasync

Python Async/Await Interface.

sequence

Python Sequence Interface Trait and support implementation for implementing sequence

Structs

PyTraverseError

PyVisit

Object visitor for GC.