Module pyo3::class

Python object protocols

Reexports

pub use self::basic::PyObjectProtocol;

pub use self::async::PyAsyncProtocol;

pub use self::iter::PyIterProtocol;

pub use self::buffer::PyBufferProtocol;

pub use self::context::PyContextProtocol;

pub use self::descr::PyDescrProtocol;

pub use self::number::PyNumberProtocol;

pub use self::mapping::PyMappingProtocol;

pub use self::sequence::PySequenceProtocol;

pub use self::gc::PyVisit;

pub use self::gc::PyGCProtocol;

pub use self::gc::PyTraverseError;

pub use self::methods::PyMethodDef;

pub use self::methods::PyMethodDefType;

pub use self::methods::PyMethodType;

pub use self::methods::PyGetterDef;

pub use self::methods::PySetterDef;

Modules

async
Python Async/Await Interface.

basic
Basic Python Object customization

buffer
Represent Python Buffer protocol implementation

context
Context manager api Trait and support implementation for context manager api

descr
Python Description Interface

gc
Python GC support

iter
Python Iterator Interface. Trait and support implementation for implementing iterators

mapping
Python Mapping Interface Trait and support implementation for implementing mapping support

methods

number
Python Number Interface Trait and support implementation for implementing number protocol

sequence
Python Sequence Interface Trait and support implementation for implementing sequence

Enums

CompareOp