pyo3

Module conversion

Defines conversions between Rust and Python types.

Traits

AsPyPointer

Returns a borrowed pointer to a Python object.

FromPyObject

Extract a type from a Python object.

FromPyObjectBound

Expected form of FromPyObject to be used in a future PyO3 release.

FromPyPointerDeprecated

Raw level conversion between *mut ffi::PyObject and PyO3 types.

IntoPy

Defines a conversion from a Rust type to a Python object.

PyTryFromDeprecated

Trait implemented by Python object types that allow a checked downcast. If T implements PyTryFrom, we can convert &PyAny to &T.

PyTryIntoDeprecated

Trait implemented by Python object types that allow a checked downcast. This trait is similar to std::convert::TryInto

ToPyObject

Conversion trait that allows various objects to be converted into PyObject.