pyo3

Module conversion

Defines conversions between Rust and Python types.

Traits

AsPyPointer

Returns a borrowed pointer to a Python object.

FromPyObject

Extract a type from a Python object.

FromPyObjectBound

Expected form of FromPyObject to be used in a future PyO3 release.

IntoPy

Defines a conversion from a Rust type to a Python object.

ToPyObject

Conversion trait that allows various objects to be converted into PyObject.