Trait pyo3::conversion::FromPyObject

pub trait FromPyObject<'py>: Sized {
    // Provided methods
    fn extract(ob: &'py PyAny) -> PyResult<Self> { ... }
    fn extract_bound(ob: &Bound<'py, PyAny>) -> PyResult<Self> { ... }
}

Extract a type from a Python object.

Normal usage is through the extract methods on Bound and Py, which forward to this trait.

Examples

use pyo3::prelude::*;
use pyo3::types::PyString;

Python::with_gil(|py| {
    // Calling `.extract()` on a `Bound` smart pointer
    let obj: Bound<'_, PyString> = PyString::new_bound(py, "blah");
    let s: String = obj.extract()?;

    // Calling `.extract(py)` on a `Py` smart pointer
    let obj: Py<PyString> = obj.unbind();
    let s: String = obj.extract(py)?;
})

During the migration of PyO3 from the “GIL Refs” API to the Bound<T> smart pointer, this trait has two methods extract and extract_bound which are defaulted to call each other. To avoid infinite recursion, implementors must implement at least one of these methods. The recommendation is to implement extract_bound and leave extract as the default implementation.

Provided Methods

fn extract(ob: &'py PyAny) -> PyResult<Self>

Extracts Self from the source GIL Ref obj.

Implementors are encouraged to implement extract_bound and leave this method as the default implementation, which will forward calls to extract_bound.

fn extract_bound(ob: &Bound<'py, PyAny>) -> PyResult<Self>

Extracts Self from the bound smart pointer obj.

Implementors are encouraged to implement this method and leave extract defaulted, as this will be most compatible with PyO3’s future API.