Trait pyo3::conversion::FromPyObject

pub trait FromPyObject<'source>: Sized {
    fn extract(ob: &'source PyAny) -> PyResult<Self>;
}

FromPyObject is implemented by various types that can be extracted from a Python object reference.

Normal usage is through the helper methods Py::extract or PyAny::extract:

let obj: Py<PyAny> = ...;
let value: &TargetType = obj.extract(py)?;

let any: &PyAny = ...;
let value: &TargetType = any.extract()?;

Note: depending on the implementation, the lifetime of the extracted result may depend on the lifetime of the obj or the prepared variable.

For example, when extracting &str from a Python byte string, the resulting string slice will point to the existing string data (lifetime: 'source). On the other hand, when extracting &str from a Python Unicode string, the preparation step will convert the string to UTF-8, and the resulting string slice will have lifetime 'prepared. Since which case applies depends on the runtime type of the Python object, both the obj and prepared variables must outlive the resulting string slice.

The trait’s conversion method takes a &PyAny argument but is called FromPyObject for historical reasons.

Required Methods

fn extract(ob: &'source PyAny) -> PyResult<Self>

Extracts Self from the source PyObject.