Trait pyo3::conversion::IntoPy

pub trait IntoPy<T>: Sized {
    fn into_py(self, py: Python<'_>) -> T;
}

Similar to std::convert::Into, just that it requires a gil token.

IntoPy<PyObject> (aka IntoPy<Py<PyAny>>) should be implemented to define a conversion from Rust to Python which can be used by most of PyO3's methods.

Required methods

fn into_py(self, py: Python<'_>) -> T

Performs the conversion.

Loading content...