Trait pyo3::conversion::PyTryFrom

pub trait PyTryFrom<'v>: Sized + PyNativeType {
    fn try_from<V: Into<&'v PyAny>>(
        value: V
    ) -> Result<&'v Self, PyDowncastError<'v>>;

    fn try_from_exact<V: Into<&'v PyAny>>(
        value: V
    ) -> Result<&'v Self, PyDowncastError<'v>>;

    unsafe fn try_from_unchecked<V: Into<&'v PyAny>>(value: V) -> &'v Self;
}

Trait implemented by Python object types that allow a checked downcast. If T implements PyTryFrom, we can convert &PyAny to &T.

This trait is similar to std::convert::TryFrom

Required methods

fn try_from<V: Into<&'v PyAny>>(
value: V
) -> Result<&'v Self, PyDowncastError<'v>>

Cast from a concrete Python object type to PyObject.

fn try_from_exact<V: Into<&'v PyAny>>(
value: V
) -> Result<&'v Self, PyDowncastError<'v>>

Cast from a concrete Python object type to PyObject. With exact type check.

unsafe fn try_from_unchecked<V: Into<&'v PyAny>>(value: V) -> &'v Self

Cast a PyAny to a specific type of PyObject. The caller must have already verified the reference is for this type.

Safety

Callers must ensure that the type is valid or risk type confusion.