Module pyo3::ffi

Raw FFI declarations for Python’s C API.

This module provides low level bindings to the Python interpreter. It is meant for advanced users only - regular PyO3 users shouldn’t need to interact with this module at all.

The contents of this module are not documented here, as it would entail basically copying the documentation from CPython. Consult the Python/C API Reference Manual for up-to-date documentation.

Safety

The functions in this module lack individual safety documentation, but generally the following apply:

Pointer arguments have to point to a valid Python object of the correct type, although null pointers are sometimes valid input.

The vast majority can only be used safely while the GIL is held.

Some functions have additional safety requirements, consult the Python/C API Reference Manual for more information.

Modules

structmember

Structs

PyByteArrayObject

PyCodeObject

PyComplexObject

PyDictObject

PyFloatObject

PyFrameObject

PyGetSetDef

PyInterpreterState

PyLongObject

PyMethodDef

PyModuleDef

PyModuleDef_Base

PyModuleDef_Slot

PyObject

PyStructSequence_Desc

PyStructSequence_Field

PyThreadState

PyTypeObject

PyType_Slot

PyType_Spec

PyVarObject

PyWeakReference

Py_complex

_node

symtable

Enums

PyGILState_STATE

Constants

MAX_CO_EXTRA_USERS

METH_CLASS

METH_COEXIST

METH_KEYWORDS

METH_NOARGS

METH_O

METH_STATIC

METH_VARARGS

PYOS_STACK_MARGIN

PYTHON_ABI_VERSION

PYTHON_API_VERSION

PY_SSIZE_T_MAX

PY_SSIZE_T_MIN

PY_STDIOTEXTMODE

PyModuleDef_HEAD_INIT

PyObject_HEAD_INIT

PySet_MINSIZE

Py_CLEANUP_SUPPORTED

Py_DTSF_ADD_DOT_0

Py_DTSF_ALT

Py_DTSF_SIGN

Py_DTST_FINITE

Py_DTST_INFINITE

Py_DTST_NAN

Py_EQ

Py_GE

Py_GT

Py_HASH_EXTERNAL

Py_HASH_FNV

Py_HASH_SIPHASH24

Py_LE

Py_LT

Py_NE

Py_PRINT_RAW

Py_TPFLAGS_BASETYPE

Set if the type allows subclassing

Py_TPFLAGS_BASE_EXC_SUBCLASS

Py_TPFLAGS_BYTES_SUBCLASS

Py_TPFLAGS_DEFAULT

Py_TPFLAGS_DICT_SUBCLASS

Py_TPFLAGS_HAVE_FINALIZE

Py_TPFLAGS_HAVE_GC

Objects support garbage collection (see objimp.h)

Py_TPFLAGS_HAVE_VERSION_TAG

This flag does nothing in Python 3.10+

Py_TPFLAGS_HEAPTYPE

Set if the type object is dynamically allocated

Py_TPFLAGS_IS_ABSTRACT

Py_TPFLAGS_LIST_SUBCLASS

Py_TPFLAGS_LONG_SUBCLASS

Py_TPFLAGS_READY

Set if the type is ‘ready’ – fully initialized

Py_TPFLAGS_READYING

Set while the type is being ‘readied’, to prevent recursive ready calls

Py_TPFLAGS_TUPLE_SUBCLASS

Py_TPFLAGS_TYPE_SUBCLASS

Py_TPFLAGS_UNICODE_SUBCLASS

Py_TPFLAGS_VALID_VERSION_TAG

Py_UNICODE_REPLACEMENT_CHARACTER

Py_am_aiter

Py_am_anext

Py_am_await

Py_bf_getbuffer

Py_bf_releasebuffer

Py_eval_input

Py_file_input

Py_mod_create

Py_mod_exec

Py_mp_ass_subscript

Py_mp_length

Py_mp_subscript

Py_nb_absolute

Py_nb_add

Py_nb_and

Py_nb_bool

Py_nb_divmod

Py_nb_float

Py_nb_floor_divide

Py_nb_index

Py_nb_inplace_add

Py_nb_inplace_and

Py_nb_inplace_floor_divide

Py_nb_inplace_lshift

Py_nb_inplace_matrix_multiply

Py_nb_inplace_multiply

Py_nb_inplace_or

Py_nb_inplace_power

Py_nb_inplace_remainder

Py_nb_inplace_rshift

Py_nb_inplace_subtract

Py_nb_inplace_true_divide

Py_nb_inplace_xor

Py_nb_int

Py_nb_invert

Py_nb_lshift

Py_nb_matrix_multiply

Py_nb_multiply

Py_nb_negative

Py_nb_or

Py_nb_positive

Py_nb_power

Py_nb_remainder

Py_nb_rshift

Py_nb_subtract

Py_nb_true_divide

Py_nb_xor

Py_single_input

Py_sq_ass_item

Py_sq_concat

Py_sq_contains

Py_sq_inplace_concat

Py_sq_inplace_repeat

Py_sq_item

Py_sq_length

Py_sq_repeat

Py_tp_alloc

Py_tp_base

Py_tp_bases

Py_tp_call

Py_tp_clear

Py_tp_dealloc

Py_tp_del

Py_tp_descr_get

Py_tp_descr_set

Py_tp_doc

Py_tp_finalize

Py_tp_free

Py_tp_getattr

Py_tp_getattro

Py_tp_getset

Py_tp_hash

Py_tp_init

Py_tp_is_gc

Py_tp_iter

Py_tp_iternext

Py_tp_members

Py_tp_methods

Py_tp_new

Py_tp_repr

Py_tp_richcompare

Py_tp_setattr

Py_tp_setattro

Py_tp_str

Py_tp_traverse

_PyHASH_MULTIPLIER

Statics

PyBaseObject_Type

built-in ‘object’

PyBool_Type

PyByteArrayIter_Type

PyByteArray_Type

PyBytesIter_Type

PyBytes_Type

PyCFunction_Type

PyCallIter_Type

PyCapsule_Type

PyClassMethodDescr_Type

PyComplex_Type

PyDictItems_Type

PyDictIterItem_Type

PyDictIterKey_Type

PyDictIterValue_Type

PyDictKeys_Type

PyDictProxy_Type

PyDictValues_Type

PyDict_Type

PyEllipsis_Type

PyEnum_Type

PyExc_ArithmeticError

PyExc_AssertionError

PyExc_AttributeError

PyExc_BaseException

PyExc_BlockingIOError

PyExc_BrokenPipeError

PyExc_BufferError

PyExc_BytesWarning

PyExc_ChildProcessError

PyExc_ConnectionAbortedError

PyExc_ConnectionError

PyExc_ConnectionRefusedError

PyExc_ConnectionResetError

PyExc_DeprecationWarning

PyExc_EOFError

PyExc_EnvironmentError

PyExc_Exception

PyExc_FileExistsError

PyExc_FileNotFoundError

PyExc_FloatingPointError

PyExc_FutureWarning

PyExc_GeneratorExit

PyExc_IOError

PyExc_ImportError

PyExc_ImportWarning

PyExc_IndentationError

PyExc_IndexError

PyExc_InterruptedError

PyExc_IsADirectoryError

PyExc_KeyError

PyExc_KeyboardInterrupt

PyExc_LookupError

PyExc_MemoryError

PyExc_ModuleNotFoundError

PyExc_NameError

PyExc_NotADirectoryError

PyExc_NotImplementedError

PyExc_OSError

PyExc_OverflowError

PyExc_PendingDeprecationWarning

PyExc_PermissionError

PyExc_ProcessLookupError

PyExc_RecursionError

PyExc_RecursionErrorInst

PyExc_ReferenceError

PyExc_ResourceWarning

PyExc_RuntimeError

PyExc_RuntimeWarning

PyExc_StopAsyncIteration

PyExc_StopIteration

PyExc_SyntaxError

PyExc_SyntaxWarning

PyExc_SystemError

PyExc_SystemExit

PyExc_TabError

PyExc_TimeoutError

PyExc_TypeError

PyExc_UnboundLocalError

PyExc_UnicodeDecodeError

PyExc_UnicodeEncodeError

PyExc_UnicodeError

PyExc_UnicodeTranslateError

PyExc_UnicodeWarning

PyExc_UserWarning

PyExc_ValueError

PyExc_Warning

PyExc_ZeroDivisionError

PyFilter_Type

PyFloat_Type

PyFrozenSet_Type

PyGetSetDescr_Type

PyListIter_Type

PyListRevIter_Type

PyList_Type

PyLongRangeIter_Type

PyLong_Type

PyMap_Type

PyMemberDescr_Type

PyMemoryView_Type

PyMethodDescr_Type

PyModuleDef_Type

PyModule_Type

PyProperty_Type

PyRangeIter_Type

PyRange_Type

PyReversed_Type

PySeqIter_Type

PySetIter_Type

PySet_Type

PySlice_Type

PySuper_Type

built-in ‘super’

PyTraceBack_Type

PyTupleIter_Type

PyTuple_Type

PyType_Type

built-in ‘type’

PyUnicodeIter_Type

PyUnicode_Type

PyWrapperDescr_Type

PyZip_Type

Py_FileSystemDefaultEncodeErrors

Py_FileSystemDefaultEncoding

Py_HasFileSystemDefaultEncoding

_PyWeakref_CallableProxyType

_PyWeakref_ProxyType

_PyWeakref_RefType

Functions

PyAnySet_Check⚠

PyAnySet_CheckExact⚠

PyArg_Parse⚠

PyArg_ParseTuple⚠

PyArg_ParseTupleAndKeywords⚠

PyArg_UnpackTuple⚠

PyArg_ValidateKeywordArguments⚠

PyBool_Check⚠

PyBool_FromLong⚠

PyByteArray_AsString⚠

PyByteArray_Check⚠

PyByteArray_CheckExact⚠

PyByteArray_Concat⚠

PyByteArray_FromObject⚠

PyByteArray_FromStringAndSize⚠

PyByteArray_Resize⚠

PyByteArray_Size⚠

PyBytes_AsString⚠

PyBytes_AsStringAndSize⚠

PyBytes_Check⚠

PyBytes_CheckExact⚠

PyBytes_Concat⚠

PyBytes_ConcatAndDel⚠

PyBytes_DecodeEscape⚠

PyBytes_FromFormat⚠

PyBytes_FromObject⚠

PyBytes_FromString⚠

PyBytes_FromStringAndSize⚠

PyBytes_Repr⚠

PyBytes_Size⚠

PyCFunction_Call⚠

PyCFunction_Check⚠

PyCFunction_ClearFreeList⚠

PyCFunction_GetFlags⚠

PyCFunction_GetFunction⚠

PyCFunction_GetSelf⚠

PyCFunction_New⚠

PyCFunction_NewEx⚠

PyCallIter_Check⚠

PyCallIter_New⚠

PyCallable_Check⚠

PyCapsule_CheckExact⚠

PyCapsule_GetContext⚠

PyCapsule_GetDestructor⚠

PyCapsule_GetName⚠

PyCapsule_GetPointer⚠

PyCapsule_Import⚠

PyCapsule_IsValid⚠

PyCapsule_New⚠

PyCapsule_SetContext⚠

PyCapsule_SetDestructor⚠

PyCapsule_SetName⚠

PyCapsule_SetPointer⚠

PyCodec_BackslashReplaceErrors⚠

PyCodec_Decode⚠

PyCodec_Decoder⚠

PyCodec_Encode⚠

PyCodec_Encoder⚠

PyCodec_IgnoreErrors⚠

PyCodec_IncrementalDecoder⚠

PyCodec_IncrementalEncoder⚠

PyCodec_KnownEncoding⚠

PyCodec_LookupError⚠

PyCodec_Register⚠

PyCodec_RegisterError⚠

PyCodec_ReplaceErrors⚠

PyCodec_StreamReader⚠

PyCodec_StreamWriter⚠

PyCodec_StrictErrors⚠

PyCodec_XMLCharRefReplaceErrors⚠

PyComplex_Check⚠

PyComplex_CheckExact⚠

PyComplex_FromDoubles⚠

PyComplex_ImagAsDouble⚠

PyComplex_RealAsDouble⚠

PyDescr_NewClassMethod⚠

PyDescr_NewGetSet⚠

PyDescr_NewMember⚠

PyDescr_NewMethod⚠

PyDictItems_Check⚠

PyDictKeys_Check⚠

PyDictProxy_New⚠

PyDictValues_Check⚠

PyDictViewSet_Check⚠

PyDict_Check⚠

PyDict_CheckExact⚠

PyDict_Clear⚠

PyDict_Contains⚠

PyDict_Copy⚠

PyDict_DelItem⚠

PyDict_DelItemString⚠

PyDict_GetItem⚠

PyDict_GetItemString⚠

PyDict_GetItemWithError⚠

PyDict_Items⚠

PyDict_Keys⚠

PyDict_Merge⚠

PyDict_MergeFromSeq2⚠

PyDict_New⚠

PyDict_Next⚠

PyDict_SetItem⚠

PyDict_SetItemString⚠

PyDict_Size⚠

PyDict_Update⚠

PyDict_Values⚠

PyErr_BadArgument⚠

PyErr_BadInternalCall⚠

PyErr_CheckSignals⚠

PyErr_Clear⚠

PyErr_Display⚠

PyErr_ExceptionMatches⚠

PyErr_Fetch⚠

PyErr_Format⚠

PyErr_GetExcInfo⚠

PyErr_GivenExceptionMatches⚠

PyErr_NewException⚠

PyErr_NewExceptionWithDoc⚠

PyErr_NoMemory⚠

PyErr_NormalizeException⚠

PyErr_Occurred⚠

PyErr_Print⚠

PyErr_PrintEx⚠

PyErr_ProgramText⚠

PyErr_ResourceWarning⚠

PyErr_Restore⚠

PyErr_SetExcInfo⚠

PyErr_SetFromErrno⚠

PyErr_SetFromErrnoWithFilename⚠

PyErr_SetFromErrnoWithFilenameObject⚠

PyErr_SetFromErrnoWithFilenameObjects⚠

PyErr_SetImportError⚠

PyErr_SetImportErrorSubclass⚠

PyErr_SetInterrupt⚠

PyErr_SetNone⚠

PyErr_SetObject⚠

PyErr_SetString⚠

PyErr_SyntaxLocation⚠

PyErr_SyntaxLocationEx⚠

PyErr_WarnEx⚠

PyErr_WarnExplicit⚠

PyErr_WarnFormat⚠

PyErr_WriteUnraisable⚠

PyEval_AcquireLock⚠

PyEval_AcquireThread⚠

PyEval_CallFunction⚠

PyEval_CallMethod⚠

PyEval_CallObject⚠

PyEval_CallObjectWithKeywords⚠

PyEval_EvalCode⚠

PyEval_EvalCodeEx⚠

PyEval_EvalFrame⚠

PyEval_EvalFrameEx⚠

PyEval_GetBuiltins⚠

PyEval_GetCallStats⚠

PyEval_GetFrame⚠

PyEval_GetFuncDesc⚠

PyEval_GetFuncName⚠

PyEval_GetGlobals⚠

PyEval_GetLocals⚠

PyEval_InitThreads⚠

PyEval_ReInitThreads⚠

PyEval_ReleaseLock⚠

PyEval_ReleaseThread⚠

PyEval_RestoreThread⚠

PyEval_SaveThread⚠

PyEval_ThreadsInitialized⚠

PyExceptionClass_Check⚠

PyExceptionInstance_Check⚠

PyExceptionInstance_Class⚠

PyException_GetCause⚠

PyException_GetContext⚠

PyException_GetTraceback⚠

PyException_SetCause⚠

PyException_SetContext⚠

PyException_SetTraceback⚠

PyFile_FromFd⚠

PyFile_GetLine⚠

PyFile_WriteObject⚠

PyFile_WriteString⚠

PyFloat_AsDouble⚠

PyFloat_Check⚠

PyFloat_CheckExact⚠

PyFloat_FromDouble⚠

PyFloat_FromString⚠

PyFloat_GetInfo⚠

PyFloat_GetMax⚠

PyFloat_GetMin⚠

PyFrame_GetLineNumber⚠

PyFrozenSet_Check⚠

PyFrozenSet_CheckExact⚠

PyFrozenSet_New⚠

PyGC_Collect⚠

PyGILState_Ensure⚠

PyGILState_GetThisThreadState⚠

PyGILState_Release⚠

PyImport_AddModule⚠

PyImport_AddModuleObject⚠

PyImport_AppendInittab⚠

PyImport_Cleanup⚠Deprecated

PyImport_ExecCodeModule⚠

PyImport_ExecCodeModuleEx⚠

PyImport_ExecCodeModuleObject⚠

PyImport_ExecCodeModuleWithPathnames⚠

PyImport_GetImporter⚠

PyImport_GetMagicNumber⚠

PyImport_GetMagicTag⚠

PyImport_GetModuleDict⚠

PyImport_Import⚠

PyImport_ImportFrozenModule⚠

PyImport_ImportFrozenModuleObject⚠

PyImport_ImportModule⚠

PyImport_ImportModuleEx⚠

PyImport_ImportModuleLevel⚠

PyImport_ImportModuleLevelObject⚠

PyImport_ImportModuleNoBlock⚠

PyImport_ReloadModule⚠

PyInterpreterState_Clear⚠

PyInterpreterState_Delete⚠

PyInterpreterState_GetID⚠

PyInterpreterState_New⚠

PyIter_Check⚠

PyIter_Next⚠

PyList_Append⚠

PyList_AsTuple⚠

PyList_Check⚠

PyList_CheckExact⚠

PyList_GetItem⚠

PyList_GetSlice⚠

PyList_Insert⚠

PyList_New⚠

PyList_Reverse⚠

PyList_SetItem⚠

PyList_SetSlice⚠

PyList_Size⚠

PyList_Sort⚠

PyLong_AsDouble⚠

PyLong_AsLong⚠

PyLong_AsLongAndOverflow⚠

PyLong_AsLongLong⚠

PyLong_AsLongLongAndOverflow⚠

PyLong_AsSize_t⚠

PyLong_AsSsize_t⚠

PyLong_AsUnsignedLong⚠

PyLong_AsUnsignedLongLong⚠

PyLong_AsUnsignedLongLongMask⚠

PyLong_AsUnsignedLongMask⚠

PyLong_AsVoidPtr⚠

PyLong_Check⚠

PyLong_CheckExact⚠

PyLong_FromDouble⚠

PyLong_FromLong⚠

PyLong_FromLongLong⚠

PyLong_FromSize_t⚠

PyLong_FromSsize_t⚠

PyLong_FromString⚠

PyLong_FromUnsignedLong⚠

PyLong_FromUnsignedLongLong⚠

PyLong_FromVoidPtr⚠

PyLong_GetInfo⚠

PyMapping_Check⚠

PyMapping_DelItem⚠

PyMapping_DelItemString⚠

PyMapping_GetItemString⚠

PyMapping_HasKey⚠

PyMapping_HasKeyString⚠

PyMapping_Items⚠

PyMapping_Keys⚠

PyMapping_Length⚠

PyMapping_SetItemString⚠

PyMapping_Size⚠

PyMapping_Values⚠

PyMarshal_ReadObjectFromString⚠

PyMarshal_WriteObjectToString⚠

PyMem_Calloc⚠

PyMem_Free⚠

PyMem_Malloc⚠

PyMem_Realloc⚠

PyMemoryView_Check⚠

PyMemoryView_FromMemory⚠

PyMemoryView_FromObject⚠

PyMemoryView_GetContiguous⚠

PyModuleDef_Init⚠

PyModule_AddFunctions⚠

PyModule_AddIntConstant⚠

PyModule_AddObject⚠

PyModule_AddStringConstant⚠

PyModule_Check⚠

PyModule_CheckExact⚠

PyModule_Create⚠

PyModule_Create2⚠

PyModule_ExecDef⚠

PyModule_FromDefAndSpec⚠

PyModule_FromDefAndSpec2⚠

PyModule_GetDef⚠

PyModule_GetDict⚠

PyModule_GetFilename⚠Deprecated

PyModule_GetFilenameObject⚠

PyModule_GetName⚠

PyModule_GetNameObject⚠

PyModule_GetState⚠

PyModule_New⚠

PyModule_NewObject⚠

PyModule_SetDocString⚠

PyNumber_Absolute⚠

PyNumber_Add⚠

PyNumber_And⚠

PyNumber_AsSsize_t⚠

PyNumber_Check⚠

PyNumber_Divmod⚠

PyNumber_Float⚠

PyNumber_FloorDivide⚠

PyNumber_InPlaceAdd⚠

PyNumber_InPlaceAnd⚠

PyNumber_InPlaceFloorDivide⚠

PyNumber_InPlaceLshift⚠

PyNumber_InPlaceMatrixMultiply⚠

PyNumber_InPlaceMultiply⚠

PyNumber_InPlaceOr⚠

PyNumber_InPlacePower⚠

PyNumber_InPlaceRemainder⚠

PyNumber_InPlaceRshift⚠

PyNumber_InPlaceSubtract⚠

PyNumber_InPlaceTrueDivide⚠

PyNumber_InPlaceXor⚠

PyNumber_Index⚠

PyNumber_Invert⚠

PyNumber_Long⚠

PyNumber_Lshift⚠

PyNumber_MatrixMultiply⚠

PyNumber_Multiply⚠

PyNumber_Negative⚠

PyNumber_Or⚠

PyNumber_Positive⚠

PyNumber_Power⚠

PyNumber_Remainder⚠

PyNumber_Rshift⚠

PyNumber_Subtract⚠

PyNumber_ToBase⚠

PyNumber_TrueDivide⚠

PyNumber_Xor⚠

PyOS_AfterFork⚠Deprecated

PyOS_AfterFork_Child⚠

PyOS_AfterFork_Parent⚠

PyOS_BeforeFork⚠

PyOS_FSPath⚠

PyOS_InitInterrupts⚠Deprecated

PyOS_InterruptOccurred⚠

PyOS_double_to_string⚠

PyOS_getsig⚠

PyOS_setsig⚠

PyOS_string_to_double⚠

PyOS_strtol⚠

PyOS_strtoul⚠

PyObject_ASCII⚠

PyObject_AsFileDescriptor⚠

PyObject_Bytes⚠

PyObject_Call⚠

PyObject_CallFunction⚠

PyObject_CallFunctionObjArgs⚠

PyObject_CallMethod⚠

PyObject_CallMethodObjArgs⚠

PyObject_CallObject⚠

PyObject_Calloc⚠

PyObject_ClearWeakRefs⚠

PyObject_DelAttr⚠

PyObject_DelAttrString⚠

PyObject_DelItem⚠

PyObject_DelItemString⚠

PyObject_Dir⚠

PyObject_Format⚠

PyObject_Free⚠

PyObject_GC_Del⚠

PyObject_GC_Track⚠

PyObject_GC_UnTrack⚠

PyObject_GenericGetAttr⚠

PyObject_GenericSetAttr⚠

PyObject_GenericSetDict⚠

PyObject_GetAttr⚠

PyObject_GetAttrString⚠

PyObject_GetItem⚠

PyObject_GetIter⚠

PyObject_HasAttr⚠

PyObject_HasAttrString⚠

PyObject_Hash⚠

PyObject_HashNotImplemented⚠

PyObject_Init⚠

PyObject_InitVar⚠

PyObject_IsInstance⚠

PyObject_IsSubclass⚠

PyObject_IsTrue⚠

PyObject_Length⚠

PyObject_Malloc⚠

PyObject_Not⚠

PyObject_Realloc⚠

PyObject_Repr⚠

PyObject_RichCompare⚠

PyObject_RichCompareBool⚠

PyObject_SelfIter⚠

PyObject_SetAttr⚠

PyObject_SetAttrString⚠

PyObject_SetItem⚠

PyObject_Size⚠

PyObject_Str⚠

PyObject_Type⚠

PyObject_TypeCheck⚠

PyRange_Check⚠

PySeqIter_Check⚠

PySeqIter_New⚠

PySequence_Check⚠

PySequence_Concat⚠

PySequence_Contains⚠

PySequence_Count⚠

PySequence_DelItem⚠

PySequence_DelSlice⚠

PySequence_Fast⚠

PySequence_GetItem⚠

PySequence_GetSlice⚠

PySequence_In⚠

PySequence_InPlaceConcat⚠

PySequence_InPlaceRepeat⚠

PySequence_Index⚠

PySequence_Length⚠

PySequence_List⚠

PySequence_Repeat⚠

PySequence_SetItem⚠

PySequence_SetSlice⚠

PySequence_Size⚠

PySequence_Tuple⚠

PySet_Add⚠

PySet_Check⚠

PySet_Clear⚠

PySet_Contains⚠

PySet_Discard⚠

PySet_New⚠

PySet_Pop⚠

PySet_Size⚠

PySlice_AdjustIndices⚠

PySlice_Check⚠

PySlice_GetIndices⚠

PySlice_GetIndicesEx⚠

PySlice_New⚠

PySlice_Unpack⚠

PyState_AddModule⚠

PyState_FindModule⚠

PyState_RemoveModule⚠

PyStructSequence_GetItem⚠

PyStructSequence_New⚠

PyStructSequence_NewType⚠

PyStructSequence_SetItem⚠

PySys_AddWarnOption⚠

PySys_AddWarnOptionUnicode⚠

PySys_AddXOption⚠

PySys_FormatStderr⚠

PySys_FormatStdout⚠

PySys_GetObject⚠

PySys_GetXOptions⚠

PySys_HasWarnOptions⚠

PySys_ResetWarnOptions⚠

PySys_SetArgv⚠

PySys_SetArgvEx⚠

PySys_SetObject⚠

PySys_SetPath⚠

PySys_WriteStderr⚠

PySys_WriteStdout⚠

PyThreadState_Clear⚠

PyThreadState_Delete⚠

PyThreadState_GET⚠

PyThreadState_Get⚠

PyThreadState_GetDict⚠

PyThreadState_New⚠

PyThreadState_SetAsyncExc⚠

PyThreadState_Swap⚠

PyTraceBack_Check⚠

PyTraceBack_Here⚠

PyTraceBack_Print⚠

PyTuple_Check⚠

PyTuple_CheckExact⚠

PyTuple_ClearFreeList⚠

PyTuple_GetItem⚠

PyTuple_GetSlice⚠

PyTuple_New⚠

PyTuple_Pack⚠

PyTuple_SetItem⚠

PyTuple_Size⚠

PyType_Check⚠

PyType_CheckExact⚠

PyType_ClearCache⚠

PyType_FastSubclass⚠

PyType_FromSpec⚠

PyType_FromSpecWithBases⚠

PyType_GenericAlloc⚠

PyType_GenericNew⚠

PyType_GetFlags⚠

PyType_GetSlot⚠

PyType_HasFeature⚠

PyType_IS_GC⚠

Test if a type has a GC head

PyType_IsSubtype⚠

PyType_Modified⚠

PyType_Ready⚠

PyUnicodeDecodeError_Create⚠

PyUnicodeDecodeError_GetEncoding⚠

PyUnicodeDecodeError_GetEnd⚠

PyUnicodeDecodeError_GetObject⚠

PyUnicodeDecodeError_GetReason⚠

PyUnicodeDecodeError_GetStart⚠

PyUnicodeDecodeError_SetEnd⚠

PyUnicodeDecodeError_SetReason⚠

PyUnicodeDecodeError_SetStart⚠

PyUnicodeEncodeError_GetEncoding⚠

PyUnicodeEncodeError_GetEnd⚠

PyUnicodeEncodeError_GetObject⚠

PyUnicodeEncodeError_GetReason⚠

PyUnicodeEncodeError_GetStart⚠

PyUnicodeEncodeError_SetEnd⚠

PyUnicodeEncodeError_SetReason⚠

PyUnicodeEncodeError_SetStart⚠

PyUnicodeTranslateError_GetEnd⚠

PyUnicodeTranslateError_GetObject⚠

PyUnicodeTranslateError_GetReason⚠

PyUnicodeTranslateError_GetStart⚠

PyUnicodeTranslateError_SetEnd⚠

PyUnicodeTranslateError_SetReason⚠

PyUnicodeTranslateError_SetStart⚠

PyUnicode_Append⚠

PyUnicode_AppendAndDel⚠

PyUnicode_AsASCIIString⚠

PyUnicode_AsCharmapString⚠

PyUnicode_AsDecodedObject⚠

PyUnicode_AsDecodedUnicode⚠

PyUnicode_AsEncodedObject⚠

PyUnicode_AsEncodedString⚠

PyUnicode_AsEncodedUnicode⚠

PyUnicode_AsLatin1String⚠

PyUnicode_AsRawUnicodeEscapeString⚠

PyUnicode_AsUCS4⚠

PyUnicode_AsUCS4Copy⚠

PyUnicode_AsUTF8String⚠

PyUnicode_AsUTF16String⚠

PyUnicode_AsUTF32String⚠

PyUnicode_AsUnicodeEscapeString⚠

PyUnicode_AsWideChar⚠

PyUnicode_AsWideCharString⚠

PyUnicode_BuildEncodingMap⚠

PyUnicode_Check⚠

PyUnicode_CheckExact⚠

PyUnicode_ClearFreeList⚠

PyUnicode_Compare⚠

PyUnicode_CompareWithASCIIString⚠

PyUnicode_Concat⚠

PyUnicode_Contains⚠

PyUnicode_Count⚠

PyUnicode_Decode⚠

PyUnicode_DecodeASCII⚠

PyUnicode_DecodeCharmap⚠

PyUnicode_DecodeFSDefault⚠

PyUnicode_DecodeFSDefaultAndSize⚠

PyUnicode_DecodeLatin1⚠

PyUnicode_DecodeLocale⚠

PyUnicode_DecodeLocaleAndSize⚠

PyUnicode_DecodeRawUnicodeEscape⚠

PyUnicode_DecodeUTF7⚠

PyUnicode_DecodeUTF7Stateful⚠

PyUnicode_DecodeUTF8⚠

PyUnicode_DecodeUTF8Stateful⚠

PyUnicode_DecodeUTF16⚠

PyUnicode_DecodeUTF16Stateful⚠

PyUnicode_DecodeUTF32⚠

PyUnicode_DecodeUTF32Stateful⚠

PyUnicode_DecodeUnicodeEscape⚠

PyUnicode_EncodeFSDefault⚠

PyUnicode_EncodeLocale⚠

PyUnicode_FSConverter⚠

PyUnicode_FSDecoder⚠

PyUnicode_Find⚠

PyUnicode_FindChar⚠

PyUnicode_Format⚠

PyUnicode_FromEncodedObject⚠

PyUnicode_FromFormat⚠

PyUnicode_FromObject⚠

PyUnicode_FromOrdinal⚠

PyUnicode_FromString⚠

PyUnicode_FromStringAndSize⚠

PyUnicode_FromWideChar⚠

PyUnicode_GetDefaultEncoding⚠

PyUnicode_GetLength⚠

PyUnicode_GetSize⚠

PyUnicode_InternFromString⚠

PyUnicode_InternImmortal⚠

PyUnicode_InternInPlace⚠

PyUnicode_IsIdentifier⚠

PyUnicode_Join⚠

PyUnicode_Partition⚠

PyUnicode_RPartition⚠

PyUnicode_RSplit⚠

PyUnicode_ReadChar⚠

PyUnicode_Replace⚠

PyUnicode_Resize⚠

PyUnicode_RichCompare⚠

PyUnicode_Split⚠

PyUnicode_Splitlines⚠

PyUnicode_Substring⚠

PyUnicode_Tailmatch⚠

PyUnicode_Translate⚠

PyUnicode_WriteChar⚠

PyWeakref_Check⚠

PyWeakref_CheckProxy⚠

PyWeakref_CheckRef⚠

PyWeakref_CheckRefExact⚠

PyWeakref_GetObject⚠

PyWeakref_NewProxy⚠

PyWeakref_NewRef⚠

PyWrapper_New⚠

Py_AddPendingCall⚠

Py_AtExit⚠

Py_BuildValue⚠

Py_BytesMain⚠

Py_CLEAR⚠

Py_CompileString⚠

Py_DECREF⚠

Py_DecRef⚠

Py_DecodeLocale⚠

Py_Ellipsis⚠

Py_EncodeLocale⚠

Py_EndInterpreter⚠

Py_Exit⚠

Py_False⚠

Py_FatalError⚠

Py_Finalize⚠

Py_FinalizeEx⚠

Py_GetBuildInfo⚠

Py_GetCompiler⚠

Py_GetCopyright⚠

Py_GetExecPrefix⚠

Py_GetPath⚠

Py_GetPlatform⚠

Py_GetPrefix⚠

Py_GetProgramFullPath⚠

Py_GetProgramName⚠

Py_GetPythonHome⚠

Py_GetRecursionLimit⚠

Py_GetVersion⚠

Py_INCREF⚠

Py_IS_TYPE⚠

Py_IncRef⚠

Py_Initialize⚠

Py_InitializeEx⚠

Py_Is⚠

Py_IsFalse⚠

Py_IsInitialized⚠

Py_IsNone⚠

Py_IsTrue⚠

Py_Main⚠

Py_MakePendingCalls⚠

Py_NewInterpreter⚠

Py_None⚠

Py_NotImplemented⚠

Py_REFCNT⚠

Py_ReprEnter⚠

Py_ReprLeave⚠

Py_SIZE⚠

Py_SetPath⚠

Py_SetProgramName⚠

Py_SetPythonHome⚠

Py_SetRecursionLimit⚠

Py_SymtableString⚠

Py_TYPE⚠

Py_True⚠

Py_XDECREF⚠

Py_XINCREF⚠

_PyErr_BadInternalCall⚠

_PyObject_GC_New⚠

_PyObject_GC_NewVar⚠

_PyObject_GC_Resize⚠

_PyObject_New⚠

_PyObject_NewVar⚠

_Py_Dealloc⚠

_Py_NewRef⚠

_Py_XNewRef⚠

Type Definitions

PyCFunction

PyCFunctionWithKeywords

PyCapsule_Destructor

Py_UCS1

Py_UCS2

Py_UCS4

Py_hash_t

Py_intptr_t

Py_ssize_t

Py_uhash_t

Py_uintptr_t

allocfunc

binaryfunc

descrgetfunc

descrsetfunc

destructor

freefunc

getattrfunc

getattrofunc

getiterfunc

getter

hashfunc

initproc

inquiry

iternextfunc

lenfunc

newfunc

objobjargproc

objobjproc

reprfunc

richcmpfunc

setattrfunc

setattrofunc

setter

ssizeargfunc

ssizeobjargproc

ssizessizeargfunc

ssizessizeobjargproc

ternaryfunc

traverseproc

unaryfunc

visitproc

Unions

PyMethodDefPointer

Function types used to implement Python callables.