Function pyo3::marshal::dumps

pub fn dumps<'a>(
    py: Python<'a>, 
    object: &impl AsPyPointer, 
    version: i32
) -> PyResult<&'a PyBytes>

Serialize an object to bytes using the Python built-in marshal module.

The built-in marshalling only supports a limited range of objects. The exact types supported depend on the version argument. The VERSION constant holds the highest version currently supported.

See the Python documentation for more details.

Example:

let dict = PyDict::new(py);
dict.set_item("aap", "noot").unwrap();
dict.set_item("mies", "wim").unwrap();
dict.set_item("zus", "jet").unwrap();

let bytes = marshal::dumps(py, dict, marshal::VERSION);