Module pyo3::prelude

PyO3’s prelude.

The purpose of this module is to alleviate imports of many commonly used items of the PyO3 crate by adding a glob import to the top of pyo3 heavy modules:

use pyo3::prelude::*;

Re-exports

pub use crate::conversion::FromPyObject;

pub use crate::conversion::IntoPy;

pub use crate::conversion::ToPyObject;

pub use crate::conversion::PyTryFrom;

Deprecated

pub use crate::conversion::PyTryInto;

Deprecated

pub use crate::marker::Python;

pub use crate::pycell::PyCell;

Deprecated

pub use crate::pycell::PyRef;

pub use crate::pycell::PyRefMut;

pub use crate::pyclass_init::PyClassInitializer;

pub use crate::types::PyAny;

pub use crate::types::PyModule;

pub use crate::PyNativeType;

Macros

wrap_pyfunction

Wraps a Rust function annotated with #[pyfunction].

wrap_pyfunction_bound

Wraps a Rust function annotated with #[pyfunction].

Structs

Borrowed

A borrowed equivalent to Bound.

Bound

A GIL-attached equivalent to Py.

Py

A GIL-independent reference to an object allocated on the Python heap.

PyErr

Represents a Python exception.

Traits

PyAnyMethods

This trait represents the Python APIs which are usable on all Python objects.

PyBoolMethods

Implementation of functionality for PyBool.

PyByteArrayMethods

Implementation of functionality for PyByteArray.

PyBytesMethods

Implementation of functionality for PyBytes.

PyCapsuleMethods

Implementation of functionality for PyCapsule.

PyComplexMethods

Implementation of functionality for PyComplex.

PyDictMethods

Implementation of functionality for PyDict.

PyFloatMethods

Implementation of functionality for PyFloat.

PyFrozenSetMethods

Implementation of functionality for PyFrozenSet.

PyListMethods

Implementation of functionality for PyList.

PyMappingMethods

Implementation of functionality for PyMapping.

PyModuleMethods

Implementation of functionality for PyModule.

PySequenceMethods

Implementation of functionality for PySequence.

PySetMethods

Implementation of functionality for PySet.

PySliceMethods

Implementation of functionality for PySlice.

PyStringMethods

Implementation of functionality for PyString.

PyTracebackMethods

Implementation of functionality for PyTraceback.

PyTupleMethods

Implementation of functionality for PyTuple.

PyTypeMethods

Implementation of functionality for PyType.

Type Definitions

PyObject

A commonly-used alias for Py<PyAny>.

PyResult

Represents the result of a Python call.

Attribute Macros

pyclass

pyfunction

A proc macro used to expose Rust functions to Python.

pymethods

A proc macro used to expose methods to Python.

pymodule

A proc macro used to implement Python modules.

Derive Macros

FromPyObject