Struct pyo3::prelude::Py

#[repr(transparent)]
pub struct Py<T>(_, _);

A Python object of known type T.

Accessing this object is thread-safe, since any access to its API requires a Python<'py> GIL token. There are a few different ways to use the Python object contained:

Py::as_ref to borrow a GIL-bound reference to the contained object.

Py::borrow, Py::try_borrow, Py::borrow_mut, or Py::try_borrow_mut, to directly access a #[pyclass] value (which has RefCell-like behavior, see the PyCell guide entry ).

Use methods directly on Py, such as Py::call and Py::call_method.

See the guide for an explanation of the different Python object types.

Implementations

impl<T> Py<T> where
T: PyClass,

pub fn new(
py: Python<'_>,
value: impl Into<PyClassInitializer<T>>
) -> PyResult<Py<T>> where
T::BaseLayout: PyBorrowFlagLayout<T::BaseType>,

Create a new instance Py<T> of a #[pyclass] on the Python heap.

impl<T> Py<T> where
T: PyTypeInfo,

pub fn as_ref<'py>(&'py self, _py: Python<'py>) -> &'py T::AsRefTarget

Borrows a GIL-bound reference to the contained T. By binding to the GIL lifetime, this allows the GIL-bound reference to not require Python for any of its methods.

For native types, this reference is &T. For pyclasses, this is &PyCell<T>.

Examples

Get access to &PyList from Py<PyList>:

let list: Py<PyList> = PyList::empty(py).into();
let list: &PyList = list.as_ref(py);
assert_eq!(list.len(), 0);

Get access to &PyCell<MyClass> from Py<MyClass>:

#[pyclass]
struct MyClass { }
let my_class: Py<MyClass> = Py::new(py, MyClass { }).unwrap();
let my_class_cell: &PyCell<MyClass> = my_class.as_ref(py);
assert!(my_class_cell.try_borrow().is_ok());

pub fn into_ref(self, py: Python<'_>) -> &T::AsRefTarget

Similar to as_ref, and also consumes this Py and registers the Python object reference in PyO3’s object storage. The reference count for the Python object will not be decreased until the GIL lifetime ends.

Example

Useful when returning GIL-bound references from functions. In the snippet below, note that the 'py lifetime of the input GIL lifetime is also given to the returned reference:

fn new_py_any<'py>(py: Python<'py>, value: impl IntoPy<PyObject>) -> &'py PyAny {
    let obj: PyObject = value.into_py(py);

    // .as_ref(py) would not be suitable here, because a reference to `obj` may not be
    // returned from the function.
    obj.into_ref(py)
}

impl<T> Py<T> where
T: PyClass,

pub fn borrow<'py>(&'py self, py: Python<'py>) -> PyRef<'py, T>

Immutably borrows the value T. This borrow lasts untill the returned PyRef exists.

Equivalent to self.as_ref(py).borrow() - see PyCell::borrow

Panics

Panics if the value is currently mutably borrowed. For a non-panicking variant, use try_borrow.

pub fn borrow_mut<'py>(&'py self, py: Python<'py>) -> PyRefMut<'py, T>

Mutably borrows the value T. This borrow lasts untill the returned PyRefMut exists.

Equivalent to self.as_ref(py).borrow_mut() - see PyCell::borrow_mut

Panics

Panics if the value is currently mutably borrowed. For a non-panicking variant, use try_borrow_mut.

pub fn try_borrow<'py>(
&'py self,
py: Python<'py>
) -> Result<PyRef<'py, T>, PyBorrowError>

Immutably borrows the value T, returning an error if the value is currently mutably borrowed. This borrow lasts untill the returned PyRef exists.

This is the non-panicking variant of borrow.

Equivalent to self.as_ref(py).try_borrow() - see PyCell::try_borrow

pub fn try_borrow_mut<'py>(
&'py self,
py: Python<'py>
) -> Result<PyRefMut<'py, T>, PyBorrowMutError>

Mutably borrows the value T, returning an error if the value is currently borrowed. This borrow lasts untill the returned PyRefMut exists.

This is the non-panicking variant of borrow_mut.

Equivalent to self.as_ref(py).try_borrow_mut() - see [PyCell::try_borrow_mut`](../pycell/struct.PyCell.html#method.try_borrow_mut)

impl<T> Py<T>

pub fn get_refcnt(&self, _py: Python<'_>) -> isize

Gets the reference count of the ffi::PyObject pointer.

pub fn clone_ref(&self, py: Python<'_>) -> Py<T>

Clones self by calling Py_INCREF() on the ptr.

pub fn is_none(&self, _py: Python<'_>) -> bool

Returns whether the object is considered to be None.

This is equivalent to the Python expression self is None.

pub fn is_true(&self, py: Python<'_>) -> PyResult<bool>

Returns whether the object is considered to be true.

This is equivalent to the Python expression bool(self).

pub fn extract<'p, D>(&'p self, py: Python<'p>) -> PyResult<D> where
D: FromPyObject<'p>,

Extracts some type from the Python object.

This is a wrapper function around FromPyObject::extract().

pub fn getattr<N>(&self, py: Python<'_>, attr_name: N) -> PyResult<PyObject> where
N: ToPyObject,

Retrieves an attribute value.

This is equivalent to the Python expression self.attr_name.

pub fn call(
&self,
py: Python<'_>,
args: impl IntoPy<Py<PyTuple>>,
kwargs: Option<&PyDict>
) -> PyResult<PyObject>

Calls the object.

This is equivalent to the Python expression self(*args, **kwargs).

pub fn call1(
&self,
py: Python<'_>,
args: impl IntoPy<Py<PyTuple>>
) -> PyResult<PyObject>

Calls the object with only positional arguments.

This is equivalent to the Python expression self(*args).

pub fn call0(&self, py: Python<'_>) -> PyResult<PyObject>

Calls the object without arguments.

This is equivalent to the Python expression self().

pub fn call_method(
&self,
py: Python<'_>,
name: &str,
args: impl IntoPy<Py<PyTuple>>,
kwargs: Option<&PyDict>
) -> PyResult<PyObject>

Calls a method on the object.

This is equivalent to the Python expression self.name(*args, **kwargs).

pub fn call_method1(
&self,
py: Python<'_>,
name: &str,
args: impl IntoPy<Py<PyTuple>>
) -> PyResult<PyObject>

Calls a method on the object with only positional arguments.

This is equivalent to the Python expression self.name(*args).

pub fn call_method0(&self, py: Python<'_>, name: &str) -> PyResult<PyObject>

Calls a method on the object with no arguments.

This is equivalent to the Python expression self.name().

pub unsafe fn from_owned_ptr(py: Python<'_>, ptr: *mut PyObject) -> Py<T>

Create a Py<T> instance by taking ownership of the given FFI pointer.

Safety

ptr must be a pointer to a Python object of type T.

Callers must own the object referred to by ptr, as this function implicitly takes ownership of that object.

Panics

Panics if ptr is null.

pub unsafe fn from_owned_ptr_or_panic(
py: Python<'_>,
ptr: *mut PyObject
) -> Py<T>

👎 Deprecated:
this is a deprecated alias for Py::from_owned_ptr

Deprecated alias for from_owned_ptr.

Safety

ptr must be a pointer to a Python object of type T.

Callers must own the object referred to by ptr, as this function implicitly takes ownership of that object.

pub unsafe fn from_owned_ptr_or_err(
py: Python<'_>,
ptr: *mut PyObject
) -> PyResult<Py<T>>

Create a Py<T> instance by taking ownership of the given FFI pointer.

If ptr is null then the current Python exception is fetched as a PyErr.

Safety

If non-null, ptr must be a pointer to a Python object of type T.

pub unsafe fn from_owned_ptr_or_opt(
_py: Python<'_>,
ptr: *mut PyObject
) -> Option<Self>

Create a Py<T> instance by taking ownership of the given FFI pointer.

If ptr is null then None is returned.

Safety

If non-null, ptr must be a pointer to a Python object of type T.

pub unsafe fn from_borrowed_ptr(py: Python<'_>, ptr: *mut PyObject) -> Py<T>

Create a Py<T> instance by creating a new reference from the given FFI pointer.

Safety

ptr must be a pointer to a Python object of type T.

Panics

Panics if ptr is null.

pub unsafe fn from_borrowed_ptr_or_err(
py: Python<'_>,
ptr: *mut PyObject
) -> PyResult<Self>

Create a Py<T> instance by creating a new reference from the given FFI pointer.

If ptr is null then the current Python exception is fetched as a PyErr.

Safety

ptr must be a pointer to a Python object of type T.

pub unsafe fn from_borrowed_ptr_or_opt(
_py: Python<'_>,
ptr: *mut PyObject
) -> Option<Self>

Create a Py<T> instance by creating a new reference from the given FFI pointer.

If ptr is null then None is returned.

Safety

ptr must be a pointer to a Python object of type T.

impl Py<PyAny>

pub fn cast_as<'p, D>(
&'p self,
py: Python<'p>
) -> Result<&'p D, PyDowncastError<'_>> where
D: PyTryFrom<'p>,

Casts the PyObject to a concrete Python object type.

This can cast only to native Python types, not types implemented in Rust. For a more flexible alternative, see Py::extract.