Struct pyo3::prelude::PyObject

#[repr(transparent)]pub struct PyObject(_);

A Python object of any type.

The Python object's lifetime is managed by Python's garbage collector, so to access the object API, a Python<'py> GIL token is required.

See the guide for an explanation of the different Python object types.

Technically, it is a safe wrapper around NonNull<ffi::PyObject>.

Implementations

impl PyObject

pub unsafe fn from_owned_ptr(_py: Python<'_>, ptr: *mut PyObject) -> PyObject

Creates a PyObject instance for the given FFI pointer. This moves ownership over the pointer into the PyObject. Undefined behavior if the pointer is NULL or invalid.

pub unsafe fn from_owned_ptr_or_panic(
py: Python<'_>,
ptr: *mut PyObject
) -> PyObject

Creates a PyObject instance for the given FFI pointer. Panics if the pointer is NULL. Undefined behavior if the pointer is invalid.

pub unsafe fn from_owned_ptr_or_err(
py: Python<'_>,
ptr: *mut PyObject
) -> PyResult<PyObject>

Constructs a PyObject from the result of a Python FFI call that returns a new reference (owned pointer). Returns Err(PyErr) if the pointer is NULL.

pub unsafe fn from_owned_ptr_or_opt(
_py: Python<'_>,
ptr: *mut PyObject
) -> Option<PyObject>

Constructs a PyObject from the result of a Python FFI call that returns a new reference (owned pointer). Returns None if the pointer is NULL.

pub unsafe fn from_borrowed_ptr(_py: Python<'_>, ptr: *mut PyObject) -> PyObject

Creates a PyObject instance for the given Python FFI pointer. Calls Py_INCREF() on the ptr. Undefined behavior if the pointer is NULL or invalid.

pub unsafe fn from_borrowed_ptr_or_err(
py: Python<'_>,
ptr: *mut PyObject
) -> PyResult<PyObject>

Creates a PyObject instance for the given Python FFI pointer. Calls Py_INCREF() on the ptr. Returns Err(PyErr) if the pointer is NULL.

pub unsafe fn from_borrowed_ptr_or_opt(
py: Python<'_>,
ptr: *mut PyObject
) -> Option<PyObject>

Creates a PyObject instance for the given Python FFI pointer. Calls Py_INCREF() on the ptr. Returns None if the pointer is NULL.

pub fn get_refcnt(&self, _py: Python<'_>) -> isize

Gets the reference count of the ffi::PyObject pointer.

pub fn clone_ref(&self, py: Python<'_>) -> Self

Clones self by calling Py_INCREF() on the ptr.

pub fn is_none(&self, _py: Python<'_>) -> bool

Returns whether the object is considered to be None.

This is equivalent to the Python expression self is None.

pub fn is_true(&self, py: Python<'_>) -> PyResult<bool>

Returns whether the object is considered to be true.

This is equivalent to the Python expression bool(self).

pub fn cast_as<'p, D>(
&'p self,
py: Python<'p>
) -> Result<&'p D, PyDowncastError> where
D: PyTryFrom<'p>,

Casts the PyObject to a concrete Python object type.

This can cast only to native Python types, not types implemented in Rust.

pub fn extract<'p, D>(&'p self, py: Python<'p>) -> PyResult<D> where
D: FromPyObject<'p>,

Extracts some type from the Python object.

This is a wrapper function around FromPyObject::extract().

pub fn getattr<N>(&self, py: Python<'_>, attr_name: N) -> PyResult<PyObject> where
N: ToPyObject,

Retrieves an attribute value.

This is equivalent to the Python expression self.attr_name.

pub fn call(
&self,
py: Python<'_>,
args: impl IntoPy<Py<PyTuple>>,
kwargs: Option<&PyDict>
) -> PyResult<PyObject>

Calls the object.

This is equivalent to the Python expression self(*args, **kwargs).

pub fn call1(
&self,
py: Python<'_>,
args: impl IntoPy<Py<PyTuple>>
) -> PyResult<PyObject>

Calls the object with only positional arguments.

This is equivalent to the Python expression self(*args).

pub fn call0(&self, py: Python<'_>) -> PyResult<PyObject>

Calls the object without arguments.

This is equivalent to the Python expression self().

pub fn call_method(
&self,
py: Python<'_>,
name: &str,
args: impl IntoPy<Py<PyTuple>>,
kwargs: Option<&PyDict>
) -> PyResult<PyObject>

Calls a method on the object.

This is equivalent to the Python expression self.name(*args, **kwargs).

pub fn call_method1(
&self,
py: Python<'_>,
name: &str,
args: impl IntoPy<Py<PyTuple>>
) -> PyResult<PyObject>

Calls a method on the object with only positional arguments.

This is equivalent to the Python expression self.name(*args).

pub fn call_method0(&self, py: Python<'_>, name: &str) -> PyResult<PyObject>

Calls a method on the object with no arguments.

This is equivalent to the Python expression self.name().