Trait pyo3::prelude::PyMappingMethods

pub trait PyMappingMethods<'py>: Sealed {
    // Required methods
    fn len(&self) -> PyResult<usize>;
    fn is_empty(&self) -> PyResult<bool>;
    fn contains<K>(&self, key: K) -> PyResult<bool>
       where K: ToPyObject;
    fn get_item<K>(&self, key: K) -> PyResult<Bound<'py, PyAny>>
       where K: ToPyObject;
    fn set_item<K, V>(&self, key: K, value: V) -> PyResult<()>
       where K: ToPyObject,
             V: ToPyObject;
    fn del_item<K>(&self, key: K) -> PyResult<()>
       where K: ToPyObject;
    fn keys(&self) -> PyResult<Bound<'py, PySequence>>;
    fn values(&self) -> PyResult<Bound<'py, PySequence>>;
    fn items(&self) -> PyResult<Bound<'py, PySequence>>;
}

Implementation of functionality for PyMapping.

These methods are defined for the Bound<'py, PyMapping> smart pointer, so to use method call syntax these methods are separated into a trait, because stable Rust does not yet support arbitrary_self_types.

Required Methods

fn len(&self) -> PyResult<usize>

Returns the number of objects in the mapping.

This is equivalent to the Python expression len(self).

fn is_empty(&self) -> PyResult<bool>

Returns whether the mapping is empty.

fn contains<K>(&self, key: K) -> PyResult<bool>
where K: ToPyObject,

Determines if the mapping contains the specified key.

This is equivalent to the Python expression key in self.

fn get_item<K>(&self, key: K) -> PyResult<Bound<'py, PyAny>>
where K: ToPyObject,

Gets the item in self with key key.

Returns an Err if the item with specified key is not found, usually KeyError.

This is equivalent to the Python expression self[key].

fn set_item<K, V>(&self, key: K, value: V) -> PyResult<()>
where K: ToPyObject, V: ToPyObject,

Sets the item in self with key key.

This is equivalent to the Python expression self[key] = value.

fn del_item<K>(&self, key: K) -> PyResult<()>
where K: ToPyObject,

Deletes the item with key key.

This is equivalent to the Python statement del self[key].

fn keys(&self) -> PyResult<Bound<'py, PySequence>>

Returns a sequence containing all keys in the mapping.

fn values(&self) -> PyResult<Bound<'py, PySequence>>

Returns a sequence containing all values in the mapping.

fn items(&self) -> PyResult<Bound<'py, PySequence>>

Returns a sequence of tuples of all (key, value) pairs in the mapping.

Object Safety

This trait is not object safe.