⚠️ Internal Docs ⚠️ Not Public API 👉 Official Docs Here

Module pyo3::proc_macro

👎 Deprecated since 0.15.0:
please import this with use pyo3::... or from the prelude instead

This is supported on crate feature macros only.

Attribute Macros

pyclass

A proc macro used to expose Rust structs and fieldless enums as Python objects.

pyfunction

A proc macro used to expose Rust functions to Python.

pymethods

A proc macro used to expose methods to Python.

pymodule

A proc macro used to implement Python modules.

pyprotopyproto

A proc macro used to implement Python’s dunder methods.