Struct pyo3::pycell::PyCell

#[repr(C)]
pub struct PyCell<T: PyClass> { /* private fields */ }

A container type for (mutably) accessing PyClass values

PyCell autodereferences to PyAny, so you can call PyAny’s methods on a PyCell<T>.

Examples

This example demonstrates getting a mutable reference of the contained PyClass.

use pyo3::prelude::*;

#[pyclass]
struct Number {
    inner: u32,
}

#[pymethods]
impl Number {
    fn increment(&mut self) {
        self.inner += 1;
    }
}

Python::with_gil(|py| {
    let n = PyCell::new(py, Number { inner: 0 })?;

    let n_mutable: &mut Number = &mut n.borrow_mut();
    n_mutable.increment();

    Ok(())
})

For more information on how, when and why (not) to use PyCell please see the module-level documentation.

Implementations

impl<T: PyClass> PyCell<T>

pub fn new(
py: Python<'_>,
value: impl Into<PyClassInitializer<T>>
) -> PyResult<&Self>

Makes a new PyCell on the Python heap and return the reference to it.

In cases where the value in the cell does not need to be accessed immediately after creation, consider Py::new as a more efficient alternative.

pub fn borrow(&self) -> PyRef<'_, T>

Immutably borrows the value T. This borrow lasts as long as the returned PyRef exists.

Panics

Panics if the value is currently mutably borrowed. For a non-panicking variant, use try_borrow.

pub fn borrow_mut(&self) -> PyRefMut<'_, T>

Mutably borrows the value T. This borrow lasts as long as the returned PyRefMut exists.

Panics

Panics if the value is currently borrowed. For a non-panicking variant, use try_borrow_mut.

pub fn try_borrow(&self) -> Result<PyRef<'_, T>, PyBorrowError>

Immutably borrows the value T, returning an error if the value is currently mutably borrowed. This borrow lasts as long as the returned PyRef exists.

This is the non-panicking variant of borrow.

Examples

#[pyclass]
struct Class {}

Python::with_gil(|py| {
    let c = PyCell::new(py, Class {}).unwrap();
    {
        let m = c.borrow_mut();
        assert!(c.try_borrow().is_err());
    }

    {
        let m = c.borrow();
        assert!(c.try_borrow().is_ok());
    }
});

pub fn try_borrow_mut(&self) -> Result<PyRefMut<'_, T>, PyBorrowMutError>

Mutably borrows the value T, returning an error if the value is currently borrowed. This borrow lasts as long as the returned PyRefMut exists.

This is the non-panicking variant of borrow_mut.

Examples

#[pyclass]
struct Class {}
Python::with_gil(|py| {
    let c = PyCell::new(py, Class {}).unwrap();
    {
        let m = c.borrow();
        assert!(c.try_borrow_mut().is_err());
    }

    assert!(c.try_borrow_mut().is_ok());
});

pub unsafe fn try_borrow_unguarded(&self) -> Result<&T, PyBorrowError>

Immutably borrows the value T, returning an error if the value is currently mutably borrowed.

Safety

This method is unsafe because it does not return a PyRef, thus leaving the borrow flag untouched. Mutably borrowing the PyCell while the reference returned by this method is alive is undefined behaviour.

Examples

#[pyclass]
struct Class {}
Python::with_gil(|py| {
    let c = PyCell::new(py, Class {}).unwrap();

    {
        let m = c.borrow_mut();
        assert!(unsafe { c.try_borrow_unguarded() }.is_err());
    }

    {
        let m = c.borrow();
        assert!(unsafe { c.try_borrow_unguarded() }.is_ok());
    }
});

pub fn replace(&self, t: T) -> T

Replaces the wrapped value with a new one, returning the old value.

Panics

Panics if the value is currently borrowed.

pub fn replace_with<F: FnOnce(&mut T) -> T>(&self, f: F) -> T

Replaces the wrapped value with a new one computed from f, returning the old value.

Panics

Panics if the value is currently borrowed.

pub fn swap(&self, other: &Self)

Swaps the wrapped value of self with the wrapped value of other.

Panics

Panics if the value in either PyCell is currently borrowed.