⚠️ Internal Docs ⚠️ Not Public API 👉 Official Docs Here

Struct pyo3::pycell::PyRefMut

pub struct PyRefMut<'p, T: PyClass> {
    inner: &'p PyCell<T>,
}

A wrapper type for a mutably borrowed value from aPyCell<T>.

See the module-level documentation for more information.

Fields
inner: &'p PyCell<T>
Implementations

impl<'p, T: PyClass> PyRefMut<'p, T>

pub fn py(&self) -> Python<'_>

Returns a Python token that is bound to the lifetime of the PyRefMut.

impl<'p, T, U> PyRefMut<'p, T> where
T: PyClass<BaseType = U>,
U: PyClass,

pub fn into_super(self) -> PyRefMut<'p, U>

Gets a PyRef<T::BaseType>.

See PyRef::into_super for more.