Struct pyo3::pyclass_init::PyClassInitializer

pub struct PyClassInitializer<T: PyClass> { /* fields omitted */ }

Initializer for our #[pyclass] system.

You can use this type to initalize complicatedly nested #[pyclass].

Example

#[pyclass(subclass)]
struct BaseClass {
    #[pyo3(get)]
    basename: &'static str,
}
#[pyclass(extends=BaseClass, subclass)]
struct SubClass {
    #[pyo3(get)]
    subname: &'static str,
}
#[pyclass(extends=SubClass)]
struct SubSubClass {
    #[pyo3(get)]
    subsubname: &'static str,
}

#[pymethods]
impl SubSubClass {
    #[new]
    fn new() -> PyClassInitializer<Self> {
        PyClassInitializer::from(BaseClass { basename: "base" })
            .add_subclass(SubClass { subname: "sub" })
            .add_subclass(SubSubClass { subsubname: "subsub" })
    }
}
let gil = Python::acquire_gil();
let py = gil.python();
let typeobj = py.get_type::<SubSubClass>();
let inst = typeobj.call((), None).unwrap();
py_run!(py, inst, r#"
        assert inst.basename == 'base'
        assert inst.subname == 'sub'
        assert inst.subsubname == 'subsub'"#);

Implementations

impl<T: PyClass> PyClassInitializer<T>

pub fn new(
init: T,
super_init: <T::BaseType as PyTypeInfo>::Initializer
) -> Self

Construct new initializer from value T and base class' initializer.

We recommend to mainly use add_subclass, instead of directly call new.

pub fn add_subclass<S>(self, subclass_value: S) -> PyClassInitializer<S> where
S: PyClass + PyTypeInfo<BaseType = T>,
S::BaseLayout: PySizedLayout<T>,
S::BaseType: PyTypeInfo<Initializer = Self>,

Constructs a new initializer from base class' initializer.

Example

#[pyclass]
struct BaseClass {
    value: u32,
}

impl BaseClass {
    fn new(value: i32) -> PyResult<Self> {
        Ok(Self {
            value: std::convert::TryFrom::try_from(value)?,
        })
    }
}
#[pyclass(extends=BaseClass)]
struct SubClass {}

#[pymethods]
impl SubClass {
    #[new]
    fn new(value: i32) -> PyResult<PyClassInitializer<Self>> {
        let base_init = PyClassInitializer::from(BaseClass::new(value)?);
        Ok(base_init.add_subclass(SubClass {}))
    }
}