Trait pyo3::pyclass_init::PyObjectInit

pub trait PyObjectInit<T: PyTypeInfo>: Sized {
    fn init_class<L: PyLayout<T>>(self, layout: &mut L);

    fn __private__(&self) -> PrivateMarker;
}

Initializer for Python types.

This trait is intended to use internally for distinguishing #[pyclass] and Python native types.

Required methods

fn init_class<L: PyLayout<T>>(self, layout: &mut L)

fn __private__(&self) -> PrivateMarker

This trait is private to implement; this method exists to make it impossible to implement outside the crate.

Loading content...