Struct pyo3::Py

#[repr(transparent)]pub struct Py<T>(_, _);

A Python object of known type.

Accessing this object is thread-safe, since any access to its API requires a Python<'py> GIL token.

See the guide for an explanation of the different Python object types.

Technically, it is a safe wrapper around NonNull<ffi::PyObject> with specified type information.

Implementations

impl<T> Py<T> where
T: PyClass,

pub fn new(
py: Python<'_>,
value: impl Into<PyClassInitializer<T>>
) -> PyResult<Py<T>> where
T::BaseLayout: PyBorrowFlagLayout<T::BaseType>,

Create a new instance Py<T> of a #[pyclass] on the Python heap.

pub fn borrow<'py>(&'py self, py: Python<'py>) -> PyRef<'py, T>

Immutably borrows the value T. This borrow lasts untill the returned PyRef exists.

Equivalent to self.as_ref(py).borrow() - see PyCell::borrow

Panics

Panics if the value is currently mutably borrowed. For a non-panicking variant, use try_borrow.

pub fn borrow_mut<'py>(&'py self, py: Python<'py>) -> PyRefMut<'py, T>

Mutably borrows the value T. This borrow lasts untill the returned PyRefMut exists.

Equivalent to self.as_ref(py).borrow_mut() - see PyCell::borrow_mut

Panics

Panics if the value is currently mutably borrowed. For a non-panicking variant, use try_borrow_mut.

pub fn try_borrow<'py>(
&'py self,
py: Python<'py>
) -> Result<PyRef<'py, T>, PyBorrowError>

Immutably borrows the value T, returning an error if the value is currently mutably borrowed. This borrow lasts untill the returned PyRef exists.

This is the non-panicking variant of borrow.

Equivalent to self.as_ref(py).try_borrow() - see PyCell::try_borrow

pub fn try_borrow_mut<'py>(
&'py self,
py: Python<'py>
) -> Result<PyRefMut<'py, T>, PyBorrowMutError>

Mutably borrows the value T, returning an error if the value is currently borrowed. This borrow lasts untill the returned PyRefMut exists.

This is the non-panicking variant of borrow_mut.

Equivalent to self.as_ref(py).try_borrow_mut() - see [PyCell::try_borrow_mut`](../pycell/struct.PyCell.html#method.try_borrow_mut)

impl<T> Py<T>

pub unsafe fn from_owned_ptr(_py: Python<'_>, ptr: *mut PyObject) -> Py<T>

Creates a Py<T> instance for the given FFI pointer.

This moves ownership over the pointer into the Py<T>. Undefined behavior if the pointer is NULL or invalid.

pub unsafe fn from_owned_ptr_or_panic(
_py: Python<'_>,
ptr: *mut PyObject
) -> Py<T>

Creates a Py<T> instance for the given FFI pointer.

Panics if the pointer is NULL. Undefined behavior if the pointer is invalid.

pub unsafe fn from_owned_ptr_or_err(
py: Python<'_>,
ptr: *mut PyObject
) -> PyResult<Py<T>>

Construct Py<T> from the result of a Python FFI call that

Returns a new reference (owned pointer). Returns Err(PyErr) if the pointer is NULL. Unsafe because the pointer might be invalid.

pub unsafe fn from_borrowed_ptr(_py: Python<'_>, ptr: *mut PyObject) -> Py<T>

Creates a Py<T> instance for the given Python FFI pointer.

Calls Py_INCREF() on the ptr. Undefined behavior if the pointer is NULL or invalid.

pub fn get_refcnt(&self, _py: Python<'_>) -> isize

Gets the reference count of the ffi::PyObject pointer.

pub fn clone_ref(&self, py: Python<'_>) -> Py<T>

Clones self by calling Py_INCREF() on the ptr.