Struct pyo3::Py

pub struct Py<T>(pub *mut PyObject, _);

Safe wrapper around unsafe *mut ffi::PyObject pointer with specified type information.

Methods

impl<T> Py<T>

unsafe fn from_owned_ptr(ptr: *mut PyObject) -> Py<T>

Creates a Py<T> instance for the given FFI pointer. This moves ownership over the pointer into the Py<T>. Undefined behavior if the pointer is NULL or invalid.

fn from_owned_ptr_or_panic(ptr: *mut PyObject) -> Py<T>

Creates a Py<T> instance for the given FFI pointer. Panics if the pointer is null. Undefined behavior if the pointer is invalid.

fn from_owned_ptr_or_err(py: Python, ptr: *mut PyObject) -> PyResult<Py<T>>

Construct Py<T> from the result of a Python FFI call that returns a new reference (owned pointer). Returns Err(PyErr) if the pointer is null. Unsafe because the pointer might be invalid.

unsafe fn from_borrowed_ptr(ptr: *mut PyObject) -> Py<T>

Creates a Py<T> instance for the given Python FFI pointer. Calls Py_INCREF() on the ptr. Undefined behavior if the pointer is NULL or invalid.

fn get_refcnt(&self) -> isize

Gets the reference count of the ffi::PyObject pointer.

fn clone_ref(&self, _py: Python) -> Py<T>

Clone self, Calls Py_INCREF() on the ptr.

impl<T> Py<T> where
T: PyTypeInfo,

fn new<F>(py: Python, f: F) -> PyResult<Py<T>> where
F: FnOnce(PyToken) -> T,
T: PyObjectAlloc<T>,

Create new instance of T and move it under python management Returns Py<T>.

fn new_ref<F>(py: Python, f: F) -> PyResult<&T> where
F: FnOnce(PyToken) -> T,
T: PyObjectAlloc<T> + PyDowncastFrom,

Create new instance of T and move it under python management. Returns references to T

fn new_mut<F>(py: Python, f: F) -> PyResult<&mut T> where
F: FnOnce(PyToken) -> T,
T: PyObjectAlloc<T> + PyDowncastFrom,

Create new instance of T and move it under python management. Returns mutable references to T