Struct pyo3::Py

#[repr(transparent)]

pub struct Py<T>(_, _);

Safe wrapper around unsafe *mut ffi::PyObject pointer with specified type information.

Py<T> is thread-safe, because any python related operations require a Python<'p> token.

Methods

impl<T> Py<T> where
T: PyTypeCreate + PyTypeObject,

pub fn new(py: Python, value: T) -> PyResult<Py<T>>

Create new instance of T and move it under python management

impl<T> Py<T>

pub unsafe fn from_owned_ptr(ptr: *mut PyObject) -> Py<T>

Creates a Py<T> instance for the given FFI pointer. This moves ownership over the pointer into the Py<T>. Undefined behavior if the pointer is NULL or invalid.

pub unsafe fn from_owned_ptr_or_panic(ptr: *mut PyObject) -> Py<T>

Creates a Py<T> instance for the given FFI pointer. Panics if the pointer is null. Undefined behavior if the pointer is invalid.

pub unsafe fn from_owned_ptr_or_err(
py: Python,
ptr: *mut PyObject
) -> PyResult<Py<T>>

Construct Py<T> from the result of a Python FFI call that returns a new reference (owned pointer). Returns Err(PyErr) if the pointer is null. Unsafe because the pointer might be invalid.

pub unsafe fn from_borrowed_ptr(ptr: *mut PyObject) -> Py<T>

Creates a Py<T> instance for the given Python FFI pointer. Calls Py_INCREF() on the ptr. Undefined behavior if the pointer is NULL or invalid.

pub fn get_refcnt(&self) -> isize

Gets the reference count of the ffi::PyObject pointer.

pub fn clone_ref(&self, _py: Python) -> Py<T>

Clone self, Calls Py_INCREF() on the ptr.