Struct pyo3::Python

pub struct Python<'p>(_);

Marker type that indicates that the GIL is currently held.

The 'Python' struct is a zero-size marker struct that is required for most Python operations. This is used to indicate that the operation accesses/modifies the Python interpreter state, and thus can only be called if the Python interpreter is initialized and the Python global interpreter lock (GIL) is acquired. The lifetime 'p represents the lifetime of the Python interpreter.

You can imagine the GIL to be a giant Mutex<PythonInterpreterState>. The type Python<'p> then acts like a reference &'p PythonInterpreterState.

Methods

impl<'p> Python<'p>

pub unsafe fn assume_gil_acquired() -> Python<'p>

Retrieve Python instance under the assumption that the GIL is already acquired at this point, and stays acquired for the lifetime 'p.

Because the output lifetime 'p is not connected to any input parameter, care must be taken that the compiler infers an appropriate lifetime for 'p when calling this function.

pub fn acquire_gil() -> GILGuard

Acquires the global interpreter lock, which allows access to the Python runtime.

If the Python runtime is not already initialized, this function will initialize it. See prepare_freethreaded_python() for details.

pub fn allow_threads<T, F>(self, f: F) -> T where
F: Send + FnOnce() -> T,

Temporarily releases the GIL, thus allowing other Python threads to run.

pub fn eval(
self,
code: &str,
globals: Option<&PyDict>,
locals: Option<&PyDict>
) -> PyResult<&'p PyObjectRef>

Evaluates a Python expression in the given context and returns the result.

If globals is None, it defaults to Python module __main__. If locals is None, it defaults to the value of globals.

pub fn run(
self,
code: &str,
globals: Option<&PyDict>,
locals: Option<&PyDict>
) -> PyResult<()>

Executes one or more Python statements in the given context.

If globals is None, it defaults to Python module __main__. If locals is None, it defaults to the value of globals.

pub fn get_type<T>(self) -> &'p PyType where
T: PyTypeObject,

Gets the Python type object for type T.

pub fn import(self, name: &str) -> PyResult<&'p PyModule>

Import the Python module with the specified name.

pub fn is_instance<T: PyTypeObject, V: ToPyPointer>(
self,
obj: &V
) -> PyResult<bool>

Check whether obj is an instance of type T like Python isinstance function

pub fn is_subclass<T, U>(self) -> PyResult<bool> where
T: PyTypeObject,
U: PyTypeObject,

Check whether type T is subclass of type U like Python issubclass function

pub fn None(self) -> PyObject

Gets the Python builtin value None.

pub fn NotImplemented(self) -> PyObject

Gets the Python builtin value NotImplemented.

impl<'p> Python<'p>

pub fn init<T, F>(self, f: F) -> PyResult<Py<T>> where
F: FnOnce(PyToken) -> T,
T: PyTypeInfo + PyObjectAlloc<T>,

Create new instance of T and move it under python management. Returns Py<T>.

pub fn init_ref<T, F>(self, f: F) -> PyResult<&'p T> where
F: FnOnce(PyToken) -> T,
T: PyTypeInfo + PyObjectAlloc<T>,

Create new instance of T and move it under python management. Created object get registered in release pool. Returns references to T

pub fn init_mut<T, F>(self, f: F) -> PyResult<&'p mut T> where
F: FnOnce(PyToken) -> T,
T: PyTypeInfo + PyObjectAlloc<T>,

Create new instance of T and move it under python management. Created object get registered in release pool. Returns mutable references to T

impl<'p> Python<'p>

pub fn checked_cast_as<T>(self, obj: PyObject) -> Result<&'p T, PyDowncastError> where
T: PyTypeInfo,

Register object in release pool, and try to downcast to specific type.

pub unsafe fn cast_as<T>(self, obj: PyObject) -> &'p T where
T: PyTypeInfo,

Register object in release pool, and do unchecked downcast to specific type.

pub unsafe fn from_borrowed_ptr_to_obj(
self,
ptr: *mut PyObject
) -> &'p PyObjectRef

Register ffi::PyObject pointer in release pool

pub unsafe fn from_owned_ptr<T>(self, ptr: *mut PyObject) -> &'p T where
T: PyTypeInfo,

Register ffi::PyObject pointer in release pool, and do unchecked downcast to specific type.

pub unsafe fn mut_from_owned_ptr<T>(self, ptr: *mut PyObject) -> &'p mut T where
T: PyTypeInfo,

Register ffi::PyObject pointer in release pool, Do unchecked downcast to specific type. Returns mutable reference.

pub unsafe fn from_owned_ptr_or_err<T>(
self,
ptr: *mut PyObject
) -> PyResult<&'p T> where
T: PyTypeInfo,

Register owned ffi::PyObject pointer in release pool. Returns Err(PyErr) if the pointer is null. do unchecked downcast to specific type.

pub unsafe fn from_owned_ptr_or_opt<T>(
self,
ptr: *mut PyObject
) -> Option<&'p T> where
T: PyTypeInfo,

Register owned ffi::PyObject pointer in release pool. Returns None if the pointer is null. do unchecked downcast to specific type.

pub unsafe fn from_borrowed_ptr<T>(self, ptr: *mut PyObject) -> &'p T where
T: PyTypeInfo,

Register borrowed ffi::PyObject pointer in release pool. Panics if the pointer is null. do unchecked downcast to specific type.

pub unsafe fn mut_from_borrowed_ptr<T>(self, ptr: *mut PyObject) -> &'p mut T where
T: PyTypeInfo,

Register borrowed ffi::PyObject pointer in release pool. Panics if the pointer is null. do unchecked downcast to specific type.

pub unsafe fn from_borrowed_ptr_or_err<T>(
self,
ptr: *mut PyObject
) -> PyResult<&'p T> where
T: PyTypeInfo,

Register borrowed ffi::PyObject pointer in release pool. Returns Err(PyErr) if the pointer is null. do unchecked downcast to specific type.

pub unsafe fn from_borrowed_ptr_or_opt<T>(
self,
ptr: *mut PyObject
) -> Option<&'p T> where
T: PyTypeInfo,

Register borrowed ffi::PyObject pointer in release pool. Returns None if the pointer is null. do unchecked downcast to specific T.

pub fn release<T>(self, ob: T) where
T: IntoPyPointer,

Release PyObject reference.

pub fn xdecref(self, ptr: *mut PyObject)

Release ffi::PyObject pointer. Undefined behavior if the pointer is invalid.