Trait pyo3::IntoPyDictPointer

pub trait IntoPyDictPointer {
    fn into_dict_ptr(self, py: Python) -> *mut PyObject;
}

Conversion trait that allows various objects to be converted into PyDict object pointer. Primary use case for this trait is call and call_method methods as keywords argument.

Required Methods

fn into_dict_ptr(self, py: Python) -> *mut PyObject

Converts self into a PyDict object pointer. Whether pointer owned or borrowed depends on implementation.