Module pyo3::type_object

Python type object information

Traits

PyBorrowFlagLayout
Marker type indicates that Self can be a base layout of PyClass.

PyLayout
T: PyLayout<U> represents that T is a concrete representaion of U in Python heap. E.g., PyCell is a concrete representaion of all pyclasses, and ffi::PyObject is of PyAny.

PySizedLayout
T: PySizedLayout<U> represents T is not a instance of PyVarObject. , in addition that T is a concrete representaion of U.

PyTypeInfo
Python type information. All Python native types(e.g., PyDict) and #[pyclass] structs implement this trait.

PyTypeObject
Python object types that have a corresponding type object.