Trait pyo3::type_object::PyLayout

pub unsafe trait PyLayout<T> { }

T: PyLayout<U> represents that T is a concrete representation of U in the Python heap. E.g., PyCell is a concrete representation of all pyclasses, and ffi::PyObject is of PyAny.

This trait is intended to be used internally.

Safety

This trait must only be implemented for types which represent valid layouts of Python objects.