Trait pyo3::type_object::PySizedLayout

pub trait PySizedLayout<T>: PyLayout<T> + Sized { }

T: PySizedLayout<U> represents that T is not a instance of PyVarObject. In addition, that T is a concrete representaion of U.