Trait pyo3::type_object::PyTypeObject

pub unsafe trait PyTypeObject {
    fn type_object() -> Py<PyType>;
}

Python object types that have a corresponding type object.

This trait is marked unsafe because not fulfilling the contract for type_object leads to UB.

See PyTypeInfo::type_object

Required methods

fn type_object() -> Py<PyType>

Returns the safe abstraction over the type object.

Loading content...