Trait pyo3::typeob::PyTypeObject

pub trait PyTypeObject {
    fn init_type();

    fn type_object() -> Py<PyType>;

    fn create(py: Python) -> PyResult<PyRawObject>
    where
        Self: Sized + PyObjectAlloc<Self> + PyTypeInfo,
    { ... }
}

Trait implemented by Python object types that have a corresponding type object.

Required Methods

fn init_type()

Initialize type object

fn type_object() -> Py<PyType>

Retrieves the type object for this Python object type.

Provided Methods

fn create(py: Python) -> PyResult<PyRawObject> where
Self: Sized + PyObjectAlloc<Self> + PyTypeInfo,

Create PyRawObject which can be initialized with rust value