Module pyo3::types

Various types defined by the Python interpreter such as int, str and tuple.

Modules

iter

Iteration over Python collections.

Structs

PyAny

Represents any Python object.

PyBool

Represents a Python bool.

PyByteArray

Represents a Python bytearray.

PyBytes

Represents a Python bytes object.

PyCFunction

Represents a builtin Python function object.

PyCapsule

Represents a Python Capsule as described in Capsules:

PyCode

Represents a Python code object.

PyComplex

Represents a Python complex object.

PyDate

Bindings around datetime.date

PyDateTime

Bindings for datetime.datetime

PyDelta

Bindings for datetime.timedelta

PyDict

Represents a Python dict.

PyDictItems

Represents a Python dict_items.

PyDictKeys

Represents a Python dict_keys.

PyDictValues

Represents a Python dict_values.

PyEllipsis

Represents the Python Ellipsis object.

PyFloat

Represents a Python float object.

PyFrame

Represents a Python frame.

PyFrozenSet

Represents a Python frozenset

PyFrozenSetBuilder

Allows building a Python frozenset one item at a time

PyFunction

Represents a Python function object.

PyInt

Represents a Python int object.

PyIterator

A Python iterator object.

PyList

Represents a Python list.

PyLong

Represents a Python int object.

PyMapping

Represents a reference to a Python object supporting the mapping protocol.

PyMemoryView

Represents a Python memoryview.

PyModule

Represents a Python module object.

PyNone

Represents the Python None object.

PyNotImplemented

Represents the Python NotImplemented object.

PySequence

Represents a reference to a Python object supporting the sequence protocol.

PySet

Represents a Python set

PySlice

Represents a Python slice.

PySliceIndices

Return value from PySlice::indices.

PyString

Represents a Python string (a Unicode string object).

PySuper

Represents a Python super object.

PyTime

Bindings for datetime.time

PyTraceback

Represents a Python traceback.

PyTuple

Represents a Python tuple object.

PyType

Represents a reference to a Python type object.

PyTzInfo

Bindings for datetime.tzinfo.

PyUnicode

Represents a Python string (a Unicode string object).

Enums

PyStringData

Represents raw data backing a Python str.

Traits

DerefToPyAny

Python objects that have a base type.

IntoPyDict

Conversion trait that allows a sequence of tuples to be converted into PyDict Primary use case for this trait is call and call_method methods as keywords argument.

PyAnyMethods

This trait represents the Python APIs which are usable on all Python objects.

PyBoolMethods

Implementation of functionality for PyBool.

PyByteArrayMethods

Implementation of functionality for PyByteArray.

PyBytesMethods

Implementation of functionality for PyBytes.

PyCapsuleMethods

Implementation of functionality for PyCapsule.

PyComplexMethods

Implementation of functionality for PyComplex.

PyDateAccess

Trait for accessing the date components of a struct containing a date.

PyDeltaAccess

Trait for accessing the components of a struct containing a timedelta.

PyDictMethods

Implementation of functionality for PyDict.

PyFloatMethods

Implementation of functionality for PyFloat.

PyFrozenSetMethods

Implementation of functionality for PyFrozenSet.

PyListMethods

Implementation of functionality for PyList.

PyMappingMethods

Implementation of functionality for PyMapping.

PyModuleMethods

Implementation of functionality for PyModule.

PySequenceMethods

Implementation of functionality for PySequence.

PySetMethods

Implementation of functionality for PySet.

PySliceMethods

Implementation of functionality for PySlice.

PyStringMethods

Implementation of functionality for PyString.

PyTimeAccess

Trait for accessing the time components of a struct containing a time.

PyTracebackMethods

Implementation of functionality for PyTraceback.

PyTupleMethods

Implementation of functionality for PyTuple.

PyTypeMethods

Implementation of functionality for PyType.

PyTzInfoAccess

Trait for accessing the components of a struct containing a tzinfo.

Functions

timezone_utcDeprecated

Deprecated form of timezone_utc_bound.

timezone_utc_bound

Equivalent to datetime.timezone.utc