Module pyo3::types

Various types defined by the Python interpreter such as int, str and tuple.

Modules

iter

Iteration over Python collections.

Structs

PyAny

Represents any Python object.

PyBool

Represents a Python bool.

PyByteArray

Represents a Python bytearray.

PyBytes

Represents a Python bytes object.

PyCFunction

Represents a builtin Python function object.

PyCapsule

Represents a Python Capsule as described in Capsules:

PyCode

Represents a Python code object.

PyComplex

Represents a Python complex object.

PyDate

Bindings around datetime.date

PyDateTime

Bindings for datetime.datetime

PyDelta

Bindings for datetime.timedelta

PyDict

Represents a Python dict.

PyDictItems

Represents a Python dict_items.

PyDictKeys

Represents a Python dict_keys.

PyDictValues

Represents a Python dict_values.

PyFloat

Represents a Python float object.

PyFrame

Represents a Python frame.

PyFrozenSet

Represents a Python frozenset

PyFrozenSetBuilder

Allows building a Python frozenset one item at a time

PyFunction

Represents a Python function object.

PyInt

Represents a Python int object.

PyIterator

A Python iterator object.

PyList

Represents a Python list.

PyLong

Represents a Python int object.

PyMapping

Represents a reference to a Python object supporting the mapping protocol.

PyModule

Represents a Python module object.

PySequence

Represents a reference to a Python object supporting the sequence protocol.

PySet

Represents a Python set

PySlice

Represents a Python slice.

PySliceIndices

Return value from PySlice::indices.

PyString

Represents a Python string (a Unicode string object).

PySuper

Represents a Python super object.

PyTime

Bindings for datetime.time

PyTraceback

Represents a Python traceback.

PyTuple

Represents a Python tuple object.

PyType

Represents a reference to a Python type object.

PyTzInfo

Bindings for datetime.tzinfo.

PyUnicode

Represents a Python string (a Unicode string object).

Enums

PyStringData

Represents raw data backing a Python str.

Traits

IntoPyDict

Conversion trait that allows a sequence of tuples to be converted into PyDict Primary use case for this trait is call and call_method methods as keywords argument.

PyDateAccess

Trait for accessing the date components of a struct containing a date.

PyDeltaAccess

Trait for accessing the components of a struct containing a timedelta.

PyTimeAccess

Trait for accessing the time components of a struct containing a time.

PyTzInfoAccess

Trait for accessing the components of a struct containing a tzinfo.

Functions

timezone_utc

Equivalent to datetime.timezone.utc