Module pyo3::types

Various types defined by the Python interpreter such as int, str and tuple.

Structs

PyAny
A Python object with GIL lifetime

PyBool
Represents a Python bool.

PyByteArray
Represents a Python bytearray.

PyBytes
Represents a Python bytes object.

PyCFunction
Represents a builtin Python function object.

PyComplex
Represents a Python complex.

PyDate
Bindings around datetime.date

PyDateTime
Bindings for datetime.datetime

PyDelta
Bindings for datetime.timedelta

PyDict
Represents a Python dict.

PyFloat
Represents a Python float object.

PyFrozenSet
Represents a Python frozenset

PyFunction
Represents a Python function object.

PyInt
Represents a Python int object.

PyIterator
A Python iterator object.

PyList
Represents a Python list.

PyLong
Represents a Python int object.

PyModule
Represents a Python module object.

PySequence
Represents a reference to a Python object supporting the sequence protocol.

PySet
Represents a Python set

PySlice
Represents a Python slice.

PySliceIndices
Represents Python slice indices.

PyString
Represents a Python string (a Unicode string object).

PyTime
Bindings for datetime.time

PyTuple
Represents a Python tuple object.

PyType
Represents a reference to a Python type object.

PyTzInfo
Bindings for datetime.tzinfo

PyUnicode
Represents a Python string (a Unicode string object).

Traits

IntoPyDict
Conversion trait that allows a sequence of tuples to be converted into PyDict Primary use case for this trait is call and call_method methods as keywords argument.

PyDateAccess
Access traits Trait for accessing the date components of a struct containing a date.

PyDeltaAccess
Trait for accessing the components of a struct containing a timedelta.

PyTimeAccess
Trait for accessing the time components of a struct containing a time.