⚠️ Internal Docs ⚠️ Not Public API 👉 Official Docs Here

Module pyo3::types::datetime

This is supported on non-Py_LIMITED_API only.

Safe Rust wrappers for types defined in the Python datetime library

For more details about these types, see the Python documentation

Macros

ffi_fun_with_autoinit

Structs

PyDate

Bindings around datetime.date

PyDateTime

Bindings for datetime.datetime

PyDelta

Bindings for datetime.timedelta

PyTime

Bindings for datetime.time

PyTzInfo

Bindings for datetime.tzinfo

Traits

PyDateAccess

Trait for accessing the date components of a struct containing a date.

PyDeltaAccess

Trait for accessing the components of a struct containing a timedelta.

PyTimeAccess

Trait for accessing the time components of a struct containing a time.

Functions

PyDateTime_Check⚠

Check if op is a PyDateTimeAPI.DateTimeType or subtype.

PyDate_Check⚠

Check if op is a PyDateTimeAPI.DateType or subtype.

PyDelta_Check⚠

Check if op is a PyDateTimeAPI.DetaType or subtype.

PyTZInfo_Check⚠

Check if op is a PyDateTimeAPI.TZInfoType or subtype.

PyTime_Check⚠

Check if op is a PyDateTimeAPI.TimeType or subtype.

ensure_datetime_api

opt_to_pyobj