Trait pyo3::types::IntoPyDict

pub trait IntoPyDict {
    fn into_py_dict(self, py: Python<'_>) -> &PyDict;
}

Conversion trait that allows a sequence of tuples to be converted into PyDict Primary use case for this trait is call and call_method methods as keywords argument.

Required methods

fn into_py_dict(self, py: Python<'_>) -> &PyDict

Converts self into a PyDict object pointer. Whether pointer owned or borrowed depends on implementation.

Loading content...