Trait pyo3::types::PyDeltaAccess

pub trait PyDeltaAccess {
    fn get_days(&self) -> i32;

    fn get_seconds(&self) -> i32;

    fn get_microseconds(&self) -> i32;
}

Trait for accessing the components of a struct containing a timedelta.

Note: These access the individual components of a (day, second, microsecond) representation of the delta, they are not intended as aliases for calculating the total duration in each of these units.

Required methods

fn get_days(&self) -> i32

fn get_seconds(&self) -> i32

fn get_microseconds(&self) -> i32