Trait pyo3::types::PyFloatMethods

pub trait PyFloatMethods<'py>: Sealed {
    // Required method
    fn value(&self) -> c_double;
}

Implementation of functionality for PyFloat.

These methods are defined for the Bound<'py, PyFloat> smart pointer, so to use method call syntax these methods are separated into a trait, because stable Rust does not yet support arbitrary_self_types.

Required Methods

fn value(&self) -> c_double

Gets the value of this float.