Trait pyo3::types::PyTimeAccess

pub trait PyTimeAccess {
    // Required methods
    fn get_hour(&self) -> u8;
    fn get_minute(&self) -> u8;
    fn get_second(&self) -> u8;
    fn get_microsecond(&self) -> u32;
    fn get_fold(&self) -> bool;
}

Trait for accessing the time components of a struct containing a time.

Required Methods

fn get_hour(&self) -> u8

Returns the hour, as an int from 0 through 23.

Implementations should conform to the upstream documentation: https://docs.python.org/3/c-api/datetime.html#c.PyDateTime_DATE_GET_HOUR

fn get_minute(&self) -> u8

Returns the minute, as an int from 0 through 59.

Implementations should conform to the upstream documentation: https://docs.python.org/3/c-api/datetime.html#c.PyDateTime_DATE_GET_MINUTE

fn get_second(&self) -> u8

Returns the second, as an int from 0 through 59.

Implementations should conform to the upstream documentation: https://docs.python.org/3/c-api/datetime.html#c.PyDateTime_DATE_GET_SECOND

fn get_microsecond(&self) -> u32

Returns the microsecond, as an int from 0 through 999999.

Implementations should conform to the upstream documentation: https://docs.python.org/3/c-api/datetime.html#c.PyDateTime_DATE_GET_MICROSECOND

fn get_fold(&self) -> bool

Returns whether this date is the later of two moments with the same representation, during a repeated interval.

This typically occurs at the end of daylight savings time. Only valid if the represented time is ambiguous. See PEP 495 for more detail.